### 2026-08-06 #11099 @editorialbot
[PRE REVIEW]: Edge Addition Planarity Suite and Generalized Graph Library
**Submitting author:** @john-boyer-phd (<a href="http://orcid.org/0000-0002-4755-5535">John Michael Boyer</a>)
**Repository:** https://github.com/graph-algorithms/edge-addition-planarity-suite
**Branch with paper.md** (empty if default branch): JOSS
**Version:** Version 5.0.0.0
