### 2026-08-06 03fe6fb
Bump TheMrMilchmann/setup-msvc-dev from 4.0.0 to 4.1.0 (#290)
Bumps [TheMrMilchmann/setup-msvc-dev](https://github.com/themrmilchmann/setup-msvc-dev) from 4.0.0 to 4.1.0.
- [Release notes](https://github.com/themrmilchmann/setup-msvc-dev/releases)
- [Commits](https://github.com/themrmilchmann/setup-msvc-dev/compare/79dac248aac9d0059f86eae9d8b5bfab4e95e97c...368ef7d1ee4d1171b31d4a7f67f4d954f903f5a9)

---
updated-dependencies:
- dependency-name: TheMrMilchmann/setup-msvc-dev
  dependency-version: 4.1.0
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-of
[…]
