### 2026-08-09 e539965
Add directed graph transpose API
