+++ c/samples/Digraph.transposeTest.txt
@@ -1,0 +1,11 @@
+N=10
+1: 5 0
+2: 1 0
+3: 2 8 0
+4: 3 0
+5: 1 4 0
+6: 1 9 0
+7: 2 10 0
+8: 6 0
+9: 7 4 0
+10: 8 5 0
