### 2026-08-09 bc6578c
Upload config.log as artifact in case of ./configure script failure
