### 2026-08-09 43592d0
Try going to very latest ubuntu to get support for -fno-sanitize-merge
Co-authored-by: Wanda B. K. Boyer <wbkboyer@gmail.com>
