### 2026-08-09 a43f84c
Remove legit NOTOK-on-NULL test due to it producing [ERROR] in debug mode (this behavior will be discontinued, but until it does, we have to avoid happy-path NOTOK occurrences).
