### 2026-08-11 c810d71
Found additional places where clean-up is required, and removed existing clean-up code in gp_WriteToString() that is no longer necessary due to sf_Free() only taking the output string from the container if the outputErrorFlag is not set before freeing theStrBuf.
g6-write-iterator.c:
* g6_SetOutputErrorFlag() is now public in g6-write-iterator.h so that the planarityApp layer and consuming applications may set the (now renamed) outputErrorFlag of the G6WriteIterator's outputContainer when non-write errors occur.
* reorganized error checking and handling in _g6_InitWriterWithStrOrFile() so that 
[…]
