+++ c/graphLib/io/g6-write-iterator.c
@@ -24,7 +24,6 @@ int _g6_WriteGraphToStrOrFile(graphP theGraph, strOrFileP *pOutputContainer);
 int _g6_InitWriterWithStrOrFile(G6WriteIteratorP theG6WriteIterator, strOrFileP *pOutputContainer);
 int _g6_InitWriter(G6WriteIteratorP theG6WriteIterator);
 int _g6_IsWriterInitialized(G6WriteIteratorP theG6WriteIterator, int reportUninitializedParts);
-void _g6_SetOutputErrorFlag(G6WriteIteratorP theG6WriteIterator);
 void _g6_PrecomputeColumnOffsets(size_t *columnOffsets, int order);
 void _g6_EncodeAdjMatAsG6(G6WriteIteratorP theG6WriteIterator);
 void _g6_GetFirstEdgeInUse(graphP theGraph, int *e, int *u, int *v);
@@ -236,9 +235,23 @@ int _g6_InitWriterWithStrOrFile(G6WriteIteratorP theG6WriteIterator, strOrFileP 
 {
     int Result = OK;
 
+    if (!sf_IsValidStrOrFile((*pOutputContainer)))
+    {
+        gp_ErrorMessage("Unable to initialize writer with invalid strOrFile "
+                        "output container.");
+        if (pOutputContainer != NULL && (*pOutputContainer) != NULL)
+        {
+            sf_SetOutputErrorFlag((*pOutputContainer));
+            sf_Free(pOutputContainer);
+        }
+        return NOTOK;
+    }
+
     if (theG6WriteIterator == NULL)
     {
         gp_ErrorMessage("Invalid parameter: theG6WriteIterator must be non-NULL.");
+        sf_SetOutputErrorFlag((*pOutputContainer));
+        sf_Free(pOutputContainer);
         return NOTOK;
     }
 
@@ -246,13 +259,9 @@ int _g6_InitWriterWithStrOrFile(G6WriteIteratorP theG6WriteIterator, strOrFileP 
     {
         gp_ErrorMessage("Unable to initialize writer, as it was already "
                         "previously initialized.");
-        return NOTOK;
-    }
-
-    if (!sf_IsValidStrOrFile((*pOutputContainer)))
-    {
-        gp_ErrorMessage("Unable to initialize writer with invalid strOrFile "
-                        "output container.");
+        sf_SetOutputErrorFlag((*pOutputContainer));
+        sf_Free(pOutputContainer);
+        g6_SetOutputErrorFlag(theG6WriteIterator);
         return NOTOK;
     }
 
@@ -264,15 +273,15 @@ int _g6_InitWriterWithStrOrFile(G6WriteIteratorP theG6WriteIterator, strOrFileP 
 
     Result = _g6_InitWriter(theG6WriteIterator);
     if (Result != OK)
-        _g6_SetOutputErrorFlag(theG6WriteIterator);
+        g6_SetOutputErrorFlag(theG6WriteIterator);
 
     return Result;
 }
 
-void _g6_SetOutputErrorFlag(G6WriteIteratorP theG6WriteIterator)
+void g6_SetOutputErrorFlag(G6WriteIteratorP theG6WriteIterator)
 {
     if (theG6WriteIterator != NULL && theG6WriteIterator->outputContainer != NULL)
-        sf_SetErrorFlag(theG6WriteIterator->outputContainer);
+        sf_SetOutputErrorFlag(theG6WriteIterator->outputContainer);
 }
 
 int _g6_InitWriter(G6WriteIteratorP theG6WriteIterator)
@@ -348,13 +357,14 @@ int g6_WriteGraph(G6WriteIteratorP theG6WriteIterator)
[…]
