+++ c/graphLib/graph.c
@@ -1908,8 +1908,6 @@ void _DetachEdgeRecord(graphP theGraph, int e)
 
 int gp_AddEdge(graphP theGraph, int u, int ulink, int v, int vlink)
 {
-    int upos, vpos;
-
     if (theGraph == NULL ||
         u < gp_LowerBoundVertexStorage(theGraph) || v < gp_LowerBoundVertexStorage(theGraph) ||
         u >= gp_UpperBoundVertexStorage(theGraph) || v >= gp_UpperBoundVertexStorage(theGraph))
@@ -1920,93 +1918,40 @@ int gp_AddEdge(graphP theGraph, int u, int ulink, int v, int vlink)
         return NOTOK;
     }
 
-    /* We enforce the edge limit */
-
-    if (gp_GetM(theGraph) >= theGraph->edgeCapacity)
-        return AT_EDGE_CAPACITY_LIMIT;
-
-    if (sp_NonEmpty(theGraph->edgeHoles))
-    {
-        sp_Pop(theGraph->edgeHoles, vpos);
-        theGraph->numEdgeHoles = sp_GetCurrentSize(theGraph->edgeHoles);
-    }
-    else
-        vpos = gp_UpperBoundEdges(theGraph);
-
-    upos = gp_GetTwin(theGraph, vpos);
-
-    gp_SetNeighbor(theGraph, upos, v);
-    _AttachEdgeRecord(theGraph, u, NIL, ulink, upos);
-    gp_SetNeighbor(theGraph, vpos, u);
-    _AttachEdgeRecord(theGraph, v, NIL, vlink, vpos);
-
-    theGraph->M++;
-    return OK;
+    return gp_InsertEdge(theGraph, u, NIL, ulink, v, NIL, vlink);
 }
 
 /********************************************************************
  gp_DynamicAddEdge()
  Refer to documentation for gp_AddEdge() for parameter description.
 
- Calls gp_AddEdge(); if AT_EDGE_CAPACITY_LIMIT, doubles the edge
- capacity using gp_EnsureEdgeCapacity(), then retries gp_AddEdge().
+ Calls gp_DynamicInsertEdge() with insertion positions that prepend or
+ append the new edge records according to ulink and vlink.
 
  Returns OK on success, NOTOK on failure.
  ********************************************************************/
 int gp_DynamicAddEdge(graphP theGraph, int u, int ulink, int v, int vlink)
 {
-    int Result = OK;
-
-    Result = gp_AddEdge(theGraph, u, ulink, v, vlink);
-
-    if (Result == AT_EDGE_CAPACITY_LIMIT)
-    {
-        // The candidate edge capacity is double the current capacity
-        int candidateEdgeCapacity = gp_GetEdgeCapacity(theGraph) << 1;
-        int N = gp_GetN(theGraph);
-        int newEdgeCapacity = candidateEdgeCapacity;
-
-        // If the candidate edge capacity exceeds the number of edges
-        // needed in an undirected clique on N vertices, then attempt
-        // to use that as the new edge capacity.
-        if (candidateEdgeCapacity > ((N * (N - 1)) >> 1))
-            newEdgeCapacity = ((N * (N - 1)) >> 1);
-
-        // However, if the edge capacity is already greater than or
-        // equal to that maximum capacity needed for an undirected
-        // clique on N vertices, then we allow the capacity to double
-        // beyond the simple undirected graph limit.
-        if (newEdgeCapacity <= gp_GetEdgeCapacity(theGraph))
-            newEdgeCapacity = candidateEdgeCapacity;
-
[…]
