### 2026-08-12 8e99563
Add basic GraphML graph writer
Signed-off-by: WilliamK112 <164879897+WilliamK112@users.noreply.github.com>
