### 2026-08-13 72324aa
test: cover GraphML output in both array modes
