### 2026-08-14 7f38042
Address GraphML review feedback
Signed-off-by: WilliamK112 <164879897+WilliamK112@users.noreply.github.com>
