### 2026-08-15 e329c55
Squashed commit of the following:
commit f25837af3c9649e11ef5bf35d62ceb19945e4fac
Author: john-boyer-phd <john.boyer.phd@gmail.com>
Date:   Sat Aug 15 09:32:10 2026 -0700

    Edit license to better align with current 3-clause BSD template

commit 1fc37eb14b7c77420bbd5cf7d048656c296c450e
Merge: 8975a84 7f38042
Author: Wanda Boyer <wbkboyer@gmail.com>
Date:   Fri Aug 14 15:13:07 2026 -0700

    Add basic GraphML graph writer (#298)

    Resolves #296

    - add `WRITE_GRAPHML` support to `gp_Write()` and `gp_WriteToString()`
    - write GraphML documents through the requested composable
    sta
[…]
