+++ c/graphLib/io/graphML-writer.c
@@ -1,0 +1,207 @@
+/*
+Copyright (c) 1997-2026, John M. Boyer
+All rights reserved.
+See the LICENSE.TXT file for licensing information.
+*/
+
+#include "graphIO.h"
+#include "strOrFile.h"
+
+/* Private functions (exported to system) */
+int _WriteGraphMLGraph(graphP theGraph, strOrFileP outputContainer);
+int _WriteGraphMLStartTag(strOrFileP outputContainer);
+int _WriteGraphMLGraphElement(graphP theGraph, int index, strOrFileP outputContainer);
+int _WriteGraphMLEndTag(strOrFileP outputContainer);
+int _WriteGraphMLGraphStartTag(graphP theGraph, int index, strOrFileP outputContainer);
+int _WriteGraphMLGraphVertices(graphP theGraph, int index, strOrFileP outputContainer);
+int _WriteGraphMLGraphEdges(graphP theGraph, int index, strOrFileP outputContainer);
+int _WriteGraphMLGraphEndTag(graphP theGraph, int index, strOrFileP outputContainer);
+
+/********************************************************************
+ _WriteGraphMLGraph()
+
+ Writes a single graph as a GraphML document.
+
+ Returns OK on success, NOTOK on failure.
+ ********************************************************************/
+int _WriteGraphMLGraph(graphP theGraph, strOrFileP outputContainer)
+{
+    if (theGraph == NULL || !sf_IsValidStrOrFile(outputContainer))
+        return NOTOK;
+
+    if (_WriteGraphMLStartTag(outputContainer) != OK ||
+        _WriteGraphMLGraphElement(theGraph, 1, outputContainer) != OK ||
+        _WriteGraphMLEndTag(outputContainer) != OK)
+        return NOTOK;
+
+    return OK;
+}
+
+/********************************************************************
+ _WriteGraphMLStartTag()
+
+ Writes the XML declaration and GraphML root start tag.
+ ********************************************************************/
+int _WriteGraphMLStartTag(strOrFileP outputContainer)
+{
+    char const *startTag =
+        "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n"
+        "<graphml xmlns=\"http://graphml.graphdrawing.org/xmlns\"\n"
+        "    xmlns:xsi=\"http://www.w3.org/2001/XMLSchema-instance\"\n"
+        "    xsi:schemaLocation=\"http://graphml.graphdrawing.org/xmlns\n"
+        "     http://graphml.graphdrawing.org/xmlns/1.0/graphml.xsd\">\n";
+
+    return sf_fputs(startTag, outputContainer) == EOF ? NOTOK : OK;
+}
+
+/********************************************************************
+ _WriteGraphMLEndTag()
+
+ Writes the GraphML root end tag.
+ ********************************************************************/
+int _WriteGraphMLEndTag(strOrFileP outputContainer)
+{
+    return sf_fputs("</graphml>\n", outputContainer) == EOF ? NOTOK : OK;
+}
+
+/********************************************************************
+ _WriteGraphMLGraphElement()
+
+ Writes a graph element and its node and edge children.
+ ********************************************************************/
+int _WriteGraphMLGraphElement(graphP theGraph, int index, strOrFileP outputContainer)
+{
+    if (theGraph == NULL || !sf_IsValidStrOrFile(outputContainer))
+        return NOTOK;
+
+    if (_WriteGraphMLGraphStartTag(theGraph, index, outputContainer) != OK ||
+        _WriteGraphMLGraphVertices(theGraph, index, outputContainer) != OK ||
[…]
