+++ LICENSE.TXT
@@ -1,26 +1,4 @@
-The Edge Addition Planarity Suite
 Copyright (c) 1997-2026, John M. Boyer
-All rights reserved. Includes a reference implementation of the following:
-
-* John M. Boyer. "Subgraph Homeomorphism via the Edge Addition Planarity Algorithm".
-  Journal of Graph Algorithms and Applications, Vol. 16, no. 2, pp. 381-410, 2012.
-  http://dx.doi.org/10.7155/jgaa.00268
-
-* John M. Boyer. "A New Method for Efficiently Generating Planar Graph
-  Visibility Representations". In P. Eades and P. Healy, editors,
-  Proceedings of the 13th International Conference on Graph Drawing 2005,
-  Lecture Notes Comput. Sci., Volume 3843, pp. 508-511, Springer-Verlag, 2006.
-  http://dx.doi.org/10.1007/11618058_47
-
-* John M. Boyer and Wendy J. Myrvold. "On the Cutting Edge: Simplified O(n)
-  Planarity by Edge Addition". Journal of Graph Algorithms and Applications,
-  Vol. 8, No. 3, pp. 241-273, 2004.
-  http://dx.doi.org/10.7155/jgaa.00091
-
-* John M. Boyer. "Simplified O(n) Algorithms for Planar Graph Embedding,
-  Kuratowski Subgraph Isolation, and Related Problems". Ph.D. Dissertation,
-  University of Victoria, 2001.
-  https://dspace.library.uvic.ca/handle/1828/9918
 
 Redistribution and use in source and binary forms, with or without
 modification, are permitted provided that the following conditions are met:

[…]
