### 2026-08-15 #11099 @john-boyer-phd
@danielskatz FYI, updated license for EAPS, and also the planarity python project, for improved 3-clause BSD template alignment. Merged to JOSS branch.
