### 2026-08-18 graph-algorithms/edge-addition-planarity-suite#301 @Deln0r
JOSS review: community guidelines files, and two README details that have gone stale
Reviewing for JOSS (openjournals/joss-reviews#11163). Three documentation points, one of which is a checklist item and two of which are small and unrelated to it.

**1. Community guidelines.** The review checklist asks for clear guidance on three things: how to contribute, how to report problems, and how to seek support.

Reporting is covered well: `.github/ISSUE_TEMPLATE/` carries a bug report and a feature request template. Contributing is covered in part, through the pull request template and the Dev Setup wiki page, though those describe the environment rather than the process. Seeking support has no stated route that I could find: Discussions are disabled, `config.yml` sets `blank_issues_enabled: false`, and no mailing list or contact is named for questions as opposed to defects.

There is no `CONTRIBUTING.md`, `CODE_OF_CONDUCT.md` or `SECURITY.md` on either `master` or the `JOSS` branch, including under `.github/`. A short `CONTRIBUTING.md` pointing at the existing wiki pages, plus a sentence naming where to ask questions, would close the item without much writing.

**2. `README.md` line 49 is missing a word.** It reads `sudo apt planarity` where line 21 correctly reads `sudo apt install planarity`.

**3. `README.md` line 21 describes the previous release asset naming.** It tells a non-developer to download `planarity-N.N.N.N.<Platform>Exe.zip`. That pattern was last used in `Version_3.0.2.0`, which shipped `planarity-3.0.2.0.WindowsExe.zip`. Since `Version_4.0.0.0` the assets have been named `planarity-N.N.N.N.MacOS-Silicon.zip` and `planarity-N.N.N.N.Windows-Intel.zip`, so a reader following the README looks for a filename that no longer exists. I verified the current one works: the 5.0.0.0 MacOS-Silicon archive runs out of the box and reproduces the expected counts on a nine-vertex input.
