### 2026-08-18 1a149cd
Readme edits for non-developer binary zip file names and for Linux package installation.
