### 2026-08-19 a64a77c
Add a code of conduct statement
