+++ CODE_OF_CONDUCT.md
@@ -1,0 +1,9 @@
+# Code of Conduct
+
+We are a community that is building something great together. So, we act professionally toward one another and are kind to each other.
+
+We are dedicated to providing a harassment-free community for everyone, regardless of gender, sexual orientation, gender identity and expression, disability, physical appearance, body size, race, or religion. We do not tolerate harassment of community members in any form. Ad hominem attacks, insults, or any other exclusionary commentary are not acceptable and will result in being blocked from further participation in this project.
+
+# How to Report
+
+Please send email to report any instance of alleged misconduct directly to the project's ombudsperson, [John M. Boyer](jboyer@acm.org), with cc to [John's backup email](john.boyer.phd@gmail.com). Your concern will be investigated and handled promptly. If you feel you have been unfairly blocked, please email the project's ombudsperson. 
