### 2026-08-19 bdc0c13
Add mailto: to email URLs
