### 2026-08-19 a93b838
Add CONTRIBUTING.md
