### 2026-08-19 11f52d6
Add a note about how to address suspected security vulnerabilities.
