### 2026-08-19 graph-algorithms/edge-addition-planarity-suite#301 @john-boyer-phd
Thank you for these excellent review comments Ian (@Deln0r). I have created a [pull request](https://github.com/graph-algorithms/edge-addition-planarity-suite/pull/302) with changes intended to address your review comments in this issue. Please could you have a look at the changes and post a comment on the PR (or here) to let me know if these changes are satisfactory?

Once you are satisfied with the changes, then we'll merge the PR to master and then sync the JOSS branch with these changes.
