### 2026-08-19 graph-algorithms/edge-addition-planarity-suite#301 @Deln0r
Thanks for the quick turnaround. Reviewed in #302: all three points are addressed, and I left two small typo notes plus one optional suggestion there.
