### 2026-08-19 #11163 @Deln0r
## Review summary

I have completed my review. The software does what the paper says it does, and I verified that at the largest scale the project itself publishes. Two checklist items remain unchecked and are explained at the end. My recommendation is to accept after those are settled.

### Correctness, verified by reproduction

The repository commits reference tables under `TestSupport/tables/` recording the expected `OK` and `NONEMBEDDABLE` counts per edge count for six algorithms. I regenerated them from scratch using the project's own pipeline (`graph_generation_orchestrator.py`, `planarity_testAllGraphs_orchestrator.py`, `test_table_generator.py`) with `geng` from Nauty 2.9.3, and compared against the committed tables.

| order | graphs | algorithm runs | result |
|---|---|---|---|
| 9 | 274,668 | 1,648,008 | every row matches |
| 10 | 12,005,168 | 72,031,008 | every row matches |
| 11 | 1,018,997,864 | 6,113,987,184 | every row matches |

That is about 6.19 billion algorithm executions, and every published count is reproduced exactly. The only differing field is the duration column, where this machine runs about twice as fast as the one that produced the committed tables.

Order 11 is worth calling out because the paper's Software design section claims validation on "more than one billion graphs having 11 or fewer vertices", and because the project's own CI cannot check it: the workflows are capped by the GitHub Actions six hour job limit. On an M3 MacBook Air the order 11 pass took 1m27s to generate 11 GB of `.g6` input and 54 minutes to run all six algorithms.

The paper's other testing claim also checks out. `c/samples/n8.mALL.g6` contains exactly 12,346 graphs, which is every graph on 8 vertices, and `planarity -test` runs all six algorithm commands over it, so the sanitiser workflow does cover all 8 vertex graphs as stated.

### Memory safety and undefined behaviour

I built with `-fsanitize=address` and with `-fsanitize=undefined -fno-sanitize-merge` on `ubuntu:26.04` with clang 21.1.8, matching `sanitisers.yml`. Both configurations build clean and pass `test-samples.sh`, and both produce zero diagnostics across the entire order 9 set, 1,648,008 runs each.

Because a silent sanitiser and a clean run look identical, I confirmed the instrumentation was live before trusting that result. The ASan executable is 2,205,336 bytes against 246,008 unsanitised, `libplanarity.so` carries 30 defined `__asan` symbols, and a fault injected into `apiutils.c` through a load-time constructor is caught on the real run path: ASan reports the heap-buffer-overflow, UBSan reports the signed overflow with file and line.

### Installation

All three documented routes work. The source build runs exactly as the README describes, with `autoreconf -fi`, `./configure`, `make dist` and `make distcheck` all succeeding and no compiler warnings across roughly 15.5k lines of C. The pre-compiled `planarity-5.0.0.0.MacOS-Silicon.zip` runs out of the box and returns counts identical to the reference on a nine vertex input. `sudo apt install planarity` works on Ubuntu 26.04 and installs 3.0.2.0, which is older, as expected for distribution packaging.

### Development history and practice

633 commits from December 2008 to August 2026, across 64 distinct active months and 14 calendar years. Both paper authors are substantive code contributors, at 492 and 86 commits. External engagement is real rather than nominal: 11 contributors besides the authors, 36 of 37 external pull requests merged, and the names include maintainers of downstream consumers such as SageMath, conda-forge and Digraphs. Six CI workflows, issue and pull request templates, dependabot, and 16 tagged releases.

### Paper

The references are complete and consistent: 13 citation keys, 13 bibliography entries, no orphans in either direction, correct pandoc syntax including multi citation and locators, and every DOI resolves. The Research impact statement is unusually well evidenced, and all twelve archived snapshots it relies on are live.

### The two unchecked items

**Community guidelines.** This is mechanical rather than substantive. #302 adds `CONTRIBUTING.md` and `CODE_OF_CONDUCT.md` and fixes the two README details from #301, and I have reviewed it and am satisfied. It targets `master` and has not yet reached the `JOSS` branch under review, so I will check the item off once it does.

**AI usage disclosure.** The section states that no generative AI tools were used in the development of this software. Two commits in the reviewed tree carry `Co-authored-by: Copilot <copilot@github.com>` and are authored by one of the paper authors: `8edae05` of 29 April, a refactor of the `testAllGraphs` flow touching 11 files, and `3deccc4` of 5 May, the MSVC variadic fix. Five merged `codex/*` branches from a non-author contributor are also in the history.

By volume this is negligible, two commits out of 633, and I want to be fair about the mechanism: GitHub inserts that trailer automatically when a maintainer accepts a Copilot suggestion during pull request review, which is not the same as generating code with a model. I am not suggesting anything was concealed. But the sentence as written is absolute, and the trailers are one `git log --grep` away for any reader, so a short qualifying clause would make the section accurate and settle the point permanently. I would rather raise it now than have it surface after publication.

### What I did not verify

The comparison against LEDA and PIGALE rests on a 2003 publication and I did not attempt to reproduce it, since LEDA is proprietary. I did not test the claim of validation on randomly generated graphs up to 10 million vertices and 30 million edges. My platform coverage was macOS on Apple silicon plus a Linux container on the same hardware, so I have no independent evidence on Windows or on x86.

There is no generated API reference, so a reader looking up a specific function reads the sources. That is an observation rather than a defect: 146 of the 163 public `gp_*` symbols carry documentation adjacent to their declaration or definition, and the algorithm entry points such as `gp_Embed` are documented thoroughly.
