### 2026-08-19 a416785
Fixing spelling and capitalization issues in `CONTRIBUTING.md`
