### 2026-08-19 #11163 @danielskatz
@Deln0r - thanks for your complete and detailed review.

Just because it is much longer and more detailed than most JOSS reviews, can you confirm that this is your personal review, not an AI-generated or performed review?
