### 2026-08-19 #11163 @danielskatz
👋 @john-boyer-phd - when you can, please address the points @Deln0r brought up that prevented him from checking off two items in his checklist.
