### 2026-08-20 #11163 @john-boyer-phd
> 👋 [@john-boyer-phd](https://github.com/john-boyer-phd) - when you can, please address the points [@Deln0r](https://github.com/Deln0r) brought up that prevented him from checking off two items in his checklist.

@danielskatz @Deln0r It’s been done, merged to master and to the JOSS branch.
