### 2026-08-20 5620615
Address review feedback on add/insert edge wrappers
Wrap delegated InsertEdge calls so NOTOK comes from the invoked API, and
rewrite the add/insert tests to early-return instead of comparing with
the DEBUG-unsafe NOTOK macro.

Signed-off-by: WilliamK112 <164879897+WilliamK112@users.noreply.github.com>
Co-authored-by: Cursor <cursoragent@cursor.com>
