### 2026-08-20 #11163 @Deln0r
Thanks, that one is done and I've checked it off. Both files are on the JOSS branch and the typos are fixed. #303 was an optional suggestion, so it doesn't gate anything.

The AI disclosure item is still open though, and I think that's my fault for burying it at the end. The paper still says no generative AI tools were used, while two commits in the reviewed tree carry a `Co-authored-by: Copilot` trailer under one of the paper's authors. A clause noting that Copilot review suggestions were accepted would close it.

If you'd rather let the sentence stand, that's a fair answer too. Say so and @danielskatz can weigh it.
