### 2026-08-20 #11163 @Deln0r
Yes, the review is mine. Every judgment in it is mine: which items I checked, the two I held and why, and the recommendation.

On the length, that was a deliberate choice. I could have posted a dozen short notes as I worked through the checklist, and instead I collected everything into one message at the end, so the thread stays readable and you get the whole picture in one place.

Since you asked, I should also be straight about how it was produced. I don't write English comfortably, so I used an AI assistant for translation and for drafting the text, and the verification runs were done with its help on my machine at my direction. What it did not do is decide anything. Reproducing the reference tables through order 11, holding the two boxes, raising the Copilot trailers: those were my calls, and I stand behind every technical claim in the review.

If that falls outside what you want for JOSS, tell me and I will redo it in whatever form you prefer.
