### 2026-08-20 98bdbb1
docs: add security reporting policy
