+++ SECURITY.md
@@ -1,0 +1,24 @@
+# Security Policy
+
+## Supported Versions
+
+This project does not currently publish a fixed support window. Please report a
+suspected vulnerability even if you observe it in an older release, and include
+the affected release tag or commit so the maintainers can assess it.
+
+## Reporting a Vulnerability
+
+Please do not disclose suspected security vulnerabilities in a public GitHub
+issue, discussion, or pull request. Instead, email the project lead,
+[John M. Boyer](mailto:jboyer@acm.org), and cc
+[John's backup email](mailto:john.boyer.phd@gmail.com).
+
+When possible, include:
+
+- the affected release, commit, platform, and configuration;
+- a description of the vulnerability and its potential impact;
+- steps or a minimal example that reproduce the issue; and
+- any known mitigations or suggested fixes.
+
+Please allow the maintainers time to investigate and coordinate a fix and any
+necessary notification to downstream consumers before public disclosure.
