### 2026-08-20 d0df165
Enable unused parameter warnings
