### 2026-08-20 00401da
Implement GraphML writer canonical metadata
