+++ c/graphLib/io/graphML-writer.c
@@ -10,12 +10,14 @@ See the LICENSE.TXT file for licensing information.
 /* Private functions (exported to system) */
 int _WriteGraphMLGraph(graphP theGraph, strOrFileP outputContainer);
 int _WriteGraphMLStartTag(strOrFileP outputContainer);
+int _WriteGraphMLKeysForGraph(graphP theGraph, strOrFileP outputContainer);
 int _WriteGraphMLGraphElement(graphP theGraph, int index, strOrFileP outputContainer);
 int _WriteGraphMLEndTag(strOrFileP outputContainer);
 int _WriteGraphMLGraphStartTag(graphP theGraph, int index, strOrFileP outputContainer);
-int _WriteGraphMLGraphVertices(graphP theGraph, int index, strOrFileP outputContainer);
-int _WriteGraphMLGraphEdges(graphP theGraph, int index, strOrFileP outputContainer);
-int _WriteGraphMLGraphEndTag(graphP theGraph, int index, strOrFileP outputContainer);
+int _WriteGraphMLGraphCustomAttributes(graphP theGraph, strOrFileP outputContainer);
+int _WriteGraphMLGraphVertices(graphP theGraph, strOrFileP outputContainer);
+int _WriteGraphMLGraphEdges(graphP theGraph, strOrFileP outputContainer);
+int _WriteGraphMLGraphEndTag(graphP theGraph, strOrFileP outputContainer);
 
 /********************************************************************
  _WriteGraphMLGraph()
@@ -30,8 +32,32 @@ int _WriteGraphMLGraph(graphP theGraph, strOrFileP outputContainer)
         return NOTOK;
 
     if (_WriteGraphMLStartTag(outputContainer) != OK ||
+        _WriteGraphMLKeysForGraph(theGraph, outputContainer) != OK ||
         _WriteGraphMLGraphElement(theGraph, 1, outputContainer) != OK ||
         _WriteGraphMLEndTag(outputContainer) != OK)
+        return NOTOK;
+
+    return OK;
+}
+
+/********************************************************************
+ _WriteGraphMLKeysForGraph()
+
+ Writes GraphML key declarations for graph-level custom attributes.
+ ********************************************************************/
+int _WriteGraphMLKeysForGraph(graphP theGraph, strOrFileP outputContainer)
+{
+    char const *zeroBasedIOKey =
+        "  <key id=\"graphflags_zerobasedio\" for=\"graph\" "
+        "attr.name=\"graphflags_zerobasedio\" attr.type=\"boolean\">\n"
+        "    <default>false</default>\n"
+        "  </key>\n";
+
+    if (theGraph == NULL || !sf_IsValidStrOrFile(outputContainer))
+        return NOTOK;
+
+    if ((gp_GetGraphFlags(theGraph) & GRAPHFLAGS_ZEROBASEDIO) &&
+        sf_fputs(zeroBasedIOKey, outputContainer) == EOF)
         return NOTOK;
 
     return OK;
@@ -75,9 +101,10 @@ int _WriteGraphMLGraphElement(graphP theGraph, int index, strOrFileP outputConta
         return NOTOK;
 
     if (_WriteGraphMLGraphStartTag(theGraph, index, outputContainer) != OK ||
-        _WriteGraphMLGraphVertices(theGraph, index, outputContainer) != OK ||
-        _WriteGraphMLGraphEdges(theGraph, index, outputContainer) != OK ||
-        _WriteGraphMLGraphEndTag(theGraph, index, outputContainer) != OK)
+        _WriteGraphMLGraphCustomAttributes(theGraph, outputContainer) != OK ||
[…]
