### 2026-08-20 #11163 @danielskatz
@john-boyer-phd - How do you respond to this?

> The AI disclosure item is still open though, and I think that's my fault for burying it at the end. The paper still says no generative AI tools were used, while two commits in the reviewed tree carry a `Co-authored-by: Copilot` trailer under one of the paper's authors. A clause noting that Copilot review suggestions were accepted would close it.
>
> If you'd rather let the sentence stand, that's a fair answer too. Say so and [@danielskatz](https://github.com/danielskatz) can weigh it.
