### 2026-08-20 #11163 @john-boyer-phd
> [@john-boyer-phd](https://github.com/john-boyer-phd) - How do you respond to this?
>
> > The AI disclosure item is still open though, and I think that's my fault for burying it at the end. The paper still says no generative AI tools were used, while two commits in the reviewed tree carry a `Co-authored-by: Copilot` trailer under one of the paper's authors. A clause noting that Copilot review suggestions were accepted would close it.
> > If you'd rather let the sentence stand, that's a fair answer too. Say so and [@danielskatz](https://github.com/danielskatz) can weigh it.

@danielskatz @Deln0r

I think there is wisdom in having me slightly amend the AI usage statement, to reflect our practices that protect the project from bad code, in light of the growing proliferation of AI tooling in current dev environments.

The two Copilot commits are false positives. VS Code changed to having Copilot as part of its core in May. It began aggressive auto completion that was unwelcome, had to be shut off, and its contributions deleted. Still, because it had edited the files with Copilot, VS Code injected Copilot as a co-author because it didn’t detect that its code was removed. For the variadic function commit, my coauthor wanted to “kick the tires” and ask its opinion, but the code it generated did not work for us at all because of the multifarious platforms we support and all the compiler warnings we ask to be errors, to maximize compatibility. Again the code was ripped out and the work done right by my coauthor, and again VS code injected Copilot as a co-author despite its code being removed. This injection is done with no clear, visual indication from _Visual_ Studio Code. Net-net, that we know of, Copilot has not contributed to this project.

For the few commits from the new contributor that have branches named starting with codex, the PRs have only the new contributor as committer, not the AI tool, so we don’t know that the tool did anything more than create the branch. That being said, whether such tools were used more extensively would only be clear to me as reviewer if there was any sloppiness in the work, which there hasn’t been.

So, it seems best for us to amend our AI usage statement to say that we’re not aware of any substantive use of AI tooling yet in this project, but in light of the growing proliferation of AI tooling, it is our review process performed without AI tools by our project owners that protects the project and its consumers from low quality or harmful contributions. As the project owners, we amend PRs or send them back for changes as needed, and we use our own expertise to ensure that the code contribution is consistent with the rest of the code in the module containing it, that it is a good solution conforming to every aspect of the technical specification in the issue, that the tests also required in the technical specification are correctly coded, that the coded solution passes the coded tests, and that all the code passes our sanitizer and compiler-warning-sensitive automated tests.

Does this sound like a reasonable amendment for me to make to the paper? If so, I can do that by Saturday and then notify you when it is done…
