### 2026-08-21 #11163 @Deln0r
@john-boyer-phd that sounds reasonable to me, and thanks for digging into where the trailers came from. The VS Code behaviour you describe explains them, and it is worth knowing that the injection happens even when the suggested code is removed, since the trailer is what a reader sees.

Your wording works. Once it is in, I will check the last item off and confirm I have finished my review.

@danielskatz for visibility.
