### 2026-08-22 dfc25e6
Update AI usage disclosure
