### 2026-08-22 ecfb06e
Commit from GitHub Actions (Generate JOSS PDF)
