### 2026-08-22 65feefa
Add directed depth-first search mode
