+++ c/samples/DirectedDFSTest.txt
@@ -1,0 +1,7 @@
+N=6
+1: 3 4 0
+2: 3 4 0
+3: 4 5 0
+4: 6 0
+5: 1 3 0
+6: 0
