### 2026-08-22 d71abda
Add statement about preferred languages for security reports.
