### 2026-08-22 c4fdbf6
Enable unused parameter warnings (#307)
Closes #299
