### 2026-08-23 eda8fb7
Fix directed DFS edge type signedness
Signed-off-by: WilliamK112 <164879897+WilliamK112@users.noreply.github.com>
