### 2026-08-23 ee30111
Complete GraphML Writer Part 2 (#308)
Closes #300

Implement GraphML writer canonical metadata

Signed-off-by: WilliamK112 <164879897+WilliamK112@users.noreply.github.com>
