### 2026-08-23 9b84536
Tweak directed DFS test to ensure dup graph preserves directed dfs flag.
