### 2026-08-23 067167a
Tweak DFSEx to directly return NOTOK in subroutine error cases (and added some linefeeds).
