### 2026-08-23 a074a2e
Add directed depth-first search mode (#309)
Closes #291

* Add directed depth-first search mode

* Fix directed DFS edge type signedness

Signed-off-by: WilliamK112 <164879897+WilliamK112@users.noreply.github.com>

* Tweak directed DFS test to ensure dup graph preserves directed dfs flag.

* Tweak DFSEx to directly return NOTOK in subroutine error cases (and added some linefeeds).

---------

Signed-off-by: WilliamK112 <164879897+WilliamK112@users.noreply.github.com>
Co-authored-by: John M. Boyer <john.boyer.phd@gmail.com>
