+++ c/graphLib/graphDFSUtils.c
@@ -16,10 +16,12 @@ See the LICENSE.TXT file for licensing information.
 #include "planarityRelated/graphPlanarity.private.h"
 
 // Private methods, except exported within library
+int _DepthFirstSearchDirected(graphP theGraph);
 int _SortVertices(graphP theGraph);
 
 // Imported methods
 extern void _ClearVertexVisitedFlags(graphP theGraph, int);
+extern int _FillVertexVisitedIndexes(graphP theGraph, int FillValue);
 
 /********************************************************************
  gp_ExtendWith_DFSUtils()
@@ -101,12 +103,6 @@ int gp_DepthFirstSearch(graphP theGraph)
     if (theGraph == NULL)
         return NOTOK;
 
-    if (gp_GetGraphFlags(theGraph) & GRAPHFLAGS_DIRECTEDEDGEDETECTED)
-    {
-        gp_ErrorMessage("gp_DepthFirstSearch() does not support directed graphs.");
-        return NOTOK;
-    }
-
     if (gp_GetGraphFlags(theGraph) & GRAPHFLAGS_DFSNUMBERED)
         return OK;
 
@@ -184,6 +180,189 @@ int gp_DepthFirstSearch(graphP theGraph)
     _gp_LogLine("graphDFSUtils.c/gp_DepthFirstSearch() end\n");
 
     theGraph->graphFlags |= GRAPHFLAGS_DFSNUMBERED;
+
+    if (gp_GetGraphFlags(theGraph) & GRAPHFLAGS_DFSNUMBERED_DIRECTED)
+    {
+        theGraph->graphFlags &= ~GRAPHFLAGS_DFSNUMBERED_DIRECTED;
+        if (_FillVertexVisitedIndexes(theGraph, 0) != OK)
+            return NOTOK;
+    }
+
+    return OK;
+}
+
+/********************************************************************
+ gp_DepthFirstSearchEx()
+
+ Routes to the requested depth-first search variant.
+ ********************************************************************/
+
+int gp_DepthFirstSearchEx(graphP theGraph, unsigned mode)
+{
+    if (theGraph == NULL)
+        return NOTOK;
+
+    if (mode == DFSMODE_UNDIRECTED)
+    {
+        if (gp_DepthFirstSearch(theGraph) == OK)
+            return OK;
+        else
+            return NOTOK;
+    }
+    else if (mode == DFSMODE_DIRECTED)
+    {
+        if (_DepthFirstSearchDirected(theGraph) == OK)
+            return OK;
+        else
+            return NOTOK;
+    }
+
+    return NOTOK;
+}
+
+/********************************************************************
+ _DepthFirstSearchDirected()
+
+ Performs a directed depth-first search. Discovery times are stored
+ in vertex index fields and finish times are stored in visitedIndex.
+ Only outgoing and undirected edge records are traversed.
+ ********************************************************************/
+
+int _DepthFirstSearchDirected(graphP theGraph)
+{
+    stackP theStack;
+    int timer = 1;
+    int v, u, w, e, stackItem, numEdgesPushed;
+
+    if (theGraph == NULL)
+        return NOTOK;
+
+    if (gp_ExtendWith_DFSUtils(theGraph) != OK)
+        return NOTOK;
+
+    theStack = theGraph->theStack;
+
+    // One integer is pushed per eligible edge record, plus at most one
+    // finish event per active DFS tree vertex. The graph stack capacity
+    // required by the undirected DFS is sufficient for these events.
+    if (sp_GetCapacity(theStack) < 2 * 2 * gp_GetM(theGraph) + 2)
+        return NOTOK;
+
+    sp_ClearStack(theStack);
+
[…]
