+++ c/planarityApp/planarityCommandLine.c
@@ -39,6 +39,7 @@ int runIdentifyContractTest(graphP theGraph);
 int runDigraphTests(void);
 int runGraphMLTests(void);
 int runDrawPlanarNonplanarWriteTest(void);
+int testDirectedDFS(void);
 int testPetersenDigraph(void);
 int testDigraphTranspose(void);
 int runGraphMLWriteTest(char const *inputFileName, char const *expectedOutputFileName);
@@ -783,7 +784,7 @@ int runFaceListTest(void)
 
 runFaceListTest_Cleanup:
 
-    if (drawing != NULL) 
+    if (drawing != NULL)
     {
         free(drawing);
         drawing = NULL;
@@ -794,7 +795,7 @@ runFaceListTest_Cleanup:
         free(faceList);
         faceList = NULL;
     }
-    
+
     gp_Free(&origGraph);
     gp_Free(&theGraph);
 
@@ -1559,6 +1560,133 @@ int callTestAllGraphs(int argc, char *argv[])
     return TestAllGraphs(commandString, infileName, outfileName, NULL);
 }
 /****************************************************************************
+ testDirectedDFS()
+ ****************************************************************************/
+
+int testDirectedDFS(void)
+{
+    graphP G = gp_New();
+    graphP G1 = NULL;
+    int const expectedDiscoveryTimes[] = {1, 11, 2, 3, 7, 4};
+    int const expectedFinishTimes[] = {10, 12, 9, 6, 8, 5};
+    int lowerVertex, v, e, source, target;
+    unsigned expectedType;
+
+    if (G == NULL)
+        return NOTOK;
+
+    if (gp_Read(G, "DirectedDFSTest.txt") != OK)
+    {
+        gp_ErrorMessage("Failed to read directed DFS sample.");
+        gp_Free(&G);
+        return NOTOK;
+    }
+
+    if (gp_DepthFirstSearchEx(G, DFSMODE_DIRECTED) != OK)
+    {
+        gp_ErrorMessage("Directed DFS returned an unexpected result.");
+        gp_Free(&G);
+        return NOTOK;
+    }
+
+    if (!(gp_GetGraphFlags(G) & GRAPHFLAGS_DFSNUMBERED_DIRECTED) ||
+        (gp_GetGraphFlags(G) & GRAPHFLAGS_DFSNUMBERED))
+    {
+        gp_ErrorMessage("Directed DFS graph flags were not set correctly.");
+        gp_Free(&G);
+        return NOTOK;
+    }
+
+    lowerVertex = gp_LowerBoundVertices(G);
+    for (v = lowerVertex; v < gp_UpperBoundVertices(G); ++v)
+    {
+        if (gp_GetIndex(G, v) != expectedDiscoveryTimes[v - lowerVertex] ||
+            gp_GetVisitedIndex(G, v) != expectedFinishTimes[v - lowerVertex])
+        {
+            gp_ErrorMessage("Directed DFS timestamp mismatch at vertex %d: "
+                            "expected (%d, %d), found (%d, %d).",
+                            v - lowerVertex + 1,
+                            expectedDiscoveryTimes[v - lowerVertex],
+                            expectedFinishTimes[v - lowerVertex],
+                            gp_GetIndex(G, v),
+                            gp_GetVisitedIndex(G, v));
+            gp_Free(&G);
+            return NOTOK;
+        }
+    }
+
+    for (e = gp_LowerBoundEdges(G); e < gp_UpperBoundEdges(G); ++e)
+    {
+        if (!gp_EdgeInUse(G, e) ||
+            gp_GetDirection(G, e) == EDGEFLAG_DIRECTION_INONLY)
+            continue;
+
[…]
