### 2026-08-28 db19f50
Propagate strOrFile read errors instead of treating them as EOF
