+++ c/tests/test-read-errors.c
@@ -1,0 +1,289 @@
+/*
+Copyright (c) 1997-2026, John M. Boyer
+All rights reserved.
+See the LICENSE.TXT file for licensing information.
+*/
+
+#if defined(HAVE_FOPENCOOKIE) && !defined(_GNU_SOURCE)
+#define _GNU_SOURCE
+#endif
+
+#include <errno.h>
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+
+#include "../graphLib/graph.h"
+#include "../graphLib/extensionSystem/graphFunctionTable.h"
+#include "../graphLib/io/g6-read-iterator.h"
+#include "../graphLib/io/strOrFile.h"
+
+extern int _ReadGraph(graphP theGraph, strOrFileP *pInputContainer);
+extern int _g6_InitReaderWithStrOrFile(G6ReadIteratorP reader, strOrFileP *input);
+extern int _g6_ReadGraphFromStrOrFile(graphP theGraph, strOrFileP *input);
+
+#if defined(HAVE_FOPENCOOKIE) || defined(HAVE_FUNOPEN)
+
+typedef struct
+{
+    const char *data;
+    size_t position;
+    int failAtEnd;
+} TestInput;
+
+static int failures = 0;
+static int checks = 0;
+static int postprocessCalls = 0;
+static const char *expectedExtraData = NULL;
+
+#define CHECK(condition)                                                        \
+    do                                                                         \
+    {                                                                          \
+        checks++;                                                              \
+        if (!(condition))                                                      \
+        {                                                                      \
+            fprintf(stderr, "Line %d: %s\n", __LINE__, #condition);              \
+            failures++;                                                        \
+        }                                                                      \
+    } while (0)
+
+#ifdef HAVE_FOPENCOOKIE
+static ssize_t readInput(void *cookie, char *buffer, size_t size)
+#else
+static int readInput(void *cookie, char *buffer, int size)
+#endif
+{
+    TestInput *input = (TestInput *)cookie;
+
+    if (size == 0)
+        return 0;
+    if (input->data[input->position] != '\0')
+    {
+        // One byte per read makes the failure boundary independent of buffering.
+        buffer[0] = input->data[input->position++];
+        return 1;
+    }
+    if (input->failAtEnd)
+    {
+        errno = EIO;
+        return -1;
+    }
+    return 0;
+}
+
+static strOrFileP newInput(TestInput *input)
+{
+    strOrFileP container = sf_NewInputContainer("placeholder", NULL);
+    FILE *stream = NULL;
+
+#ifdef HAVE_FOPENCOOKIE
+    cookie_io_functions_t functions = {readInput, NULL, NULL, NULL};
+    stream = fopencookie(input, "r", functions);
+#else
+    stream = funopen(input, readInput, NULL, NULL, NULL);
+#endif
+    if (container == NULL || stream == NULL)
+    {
+        fprintf(stderr, "Unable to create test input stream.\n");
+        exit(EXIT_FAILURE);
+    }
+    sb_Free(&container->theStrBuf);
+    container->pFile = stream;
+    return container;
+}
+
+static graphP newGraph(void)
+{
+    graphP graph = gp_New();
+    if (graph == NULL)
+        exit(EXIT_FAILURE);
+    return graph;
+}
+
+static void testLineReads(void)
+{
[…]
