### 2026-08-28 graph-algorithms/edge-addition-planarity-suite#315 @MODSetter
Signed integer overflow in random graph generation: `planarity -rm N` fails for N >= 46342
`gp_CreateRandomGraphEx()` computes the maximum simple graph edge count in a 32-bit `int`, and the
multiplication happens before the shift, so `N * (N - 1)` has to fit rather than the halved value.

`c/graphLib/graph.c:1333` (master `36762b5`):

```c
    int N, maxNumEdges, maxPlanarEdges, numPlanarCoreEdges;
    ...
    maxNumEdges = (N * (N - 1)) >> 1;
```

The largest safe vertex count is 46341, because `46341 * 46340 = 2,147,441,940` fits in an `int` and
`46342 * 46341 = 2,147,534,622` does not. Above that the product is undefined behaviour and in
practice wraps. When it wraps negative, the clamp two lines later sets `numEdges` negative, the
`numEdges < N - 1` guard returns `NOTOK`, and the command fails outright:

```
$ planarity -rm 46342 /tmp/out.g6
Creating the random graph...
[ERROR] gp_CreateRandomGraphEx() failed
	on line 596 of 'c/planarityApp/planarityRandomGraphs.c'
$ echo $?
255
```

No output file is written. Whether a given N works depends only on the sign of the wrapped value, so
the failures are not contiguous:

| N | `N*(N-1)` | as int32 | `maxNumEdges` | `planarity -rm N` |
|---|---|---|---|---|
| 1,000 | 999,000 | fits | 499,500 | works |
| 46,341 | 2,147,441,940 | fits | 1,073,720,970 | works, last good N |
| 46,342 | 2,147,534,622 | -2,147,432,674 | -1,073,716,337 | fails, exit 255 |
| 46,343 | 2,147,627,306 | -2,147,339,990 | -1,073,669,995 | fails, exit 255 |
| 65,536 | 4,294,901,760 | -65,536 | -32,768 | fails, exit 255 |
| 100,000 | 9,999,900,000 | +1,409,965,408 | +704,982,704 | works |
| 1,000,000 | 999,999,000,000 | -728,379,968 | -364,189,984 | fails, exit 255 |
| 10,000,000 | 99,999,990,000,000 | +266,447,232 | +133,223,616 | works |

A UBSan build confirms the cause:

```
c/graphLib/graph.c: runtime error: signed integer overflow: 46342 * 46341 cannot be represented in type 'int'
c/graphLib/graph.c: runtime error: signed integer overflow: 1000000 * 999999 cannot be represented in type 'int'
c/graphLib/graph.c: runtime error: signed integer overflow: 10000000 * 9999999 cannot be represented in type 'int'
```

It is not specific to one platform or word size. The boundary sits at the same N on every target I
tried, which is what you would expect since it depends on `int` being 32 bits rather than on the
pointer width:

| Platform | compiler | `-rm 46341` | `-rm 46342` |
|---|---|---|---|
| Ubuntu 24.04 x86_64 | gcc 13.3.0 | exit 0 | exit 255 |
| Ubuntu 24.04 aarch64 | gcc 13.3.0 | exit 0 | exit 255 |
| Windows 11 x64 | MSVC 19.44 | exit 0 | exit -1 |
| Windows 11 x86 | MSVC 19.44 | exit 0 | exit -1 |

## Two related things in the same area

The same expression appears at two other places, with the same 46342 ceiling:

```
c/graphLib/graph.c:1229-1230   if (M > N * (N - 1) / 2)  /  M = N * (N - 1) / 2;   gp_CreateRandomGraph()
c/graphLib/graph.c:1978-1979   if (candidateEdgeCapacity > ((N * (N - 1)) >> 1))   gp_DynamicInsertEdge()
```

UBSan also flags `gp_GetRandomNumber()` in `c/graphLib/lowLevelUtils/apiutils.c`, where `N` starts at
`rand()` and can already be near `INT_MAX` before the folding adds to it:

```
c/graphLib/lowLevelUtils/apiutils.c:101:7: runtime error: signed integer overflow: 255 + 2147483516 cannot be represented in type 'int'
```

Line 100 has the same shape (`N += ((N & 0xFFFF0000) >> 16);`) though I did not observe it trip. The
mask to `0x7FFFFFFF` on line 102 looks like it was meant to keep the value in range, but it runs after
the additions rather than before them.

## Why CI does not catch it

`sanitisers.yml` runs ASan and UBSan only over `c/samples`, which is graphs on 8 vertices. `N` there is
four orders of magnitude below the threshold, so the job that would have caught this instantly cannot
reach the code path. Adding a single large N case would cover the whole class, and it is fast:
`planarity -rm 46342` under UBSan takes under a second.

## Suggested fix

Widen the arithmetic, or guard N. Something like:

```c
    long long maxNumEdgesLL = ((long long)N * (N - 1)) >> 1;
    maxNumEdges = maxNumEdgesLL > INT_MAX ? INT_MAX : (int)maxNumEdgesLL;
```

and the same at the other two sites. For `gp_GetRandomNumber()`, applying the `0x7FFFFFFF` mask to
`rand()` before the additions, or doing the folding in `unsigned`, removes the overflow.

## Where this came from

This is not long-standing code. `git log -S` puts the expression in `a785616` (2026-07-13), which
reached master through #279 ("Improve `gp_CreateRandomGraphEx` generation"). I mention it only because
it narrows the regression window, and because the test matrix in that PR tops out at N = 20, so
nothing in it goes near the threshold.

## Reproducing

```bash
git clone https://github.com/graph-algorithms/edge-addition-planarity-suite.git
cd edge-addition-planarity-suite
autoreconf -fi && ./configure && make

./planarity -rm 46341 /tmp/a.g6 ; echo "46341 -> $?"
./planarity -rm 46342 /tmp/b.g6 ; echo "46342 -> $?"

# under UBSan
make distclean
./configure CFLAGS="-O1 -g -fsanitize=undefined,signed-integer-overflow -fno-omit-frame-pointer"
make
UBSAN_OPTIONS=print_stacktrace=1 ./planarity -rm 46342 /tmp/c.g6
```

Verified on Ubuntu 24.04, gcc 13.3.0, x86_64, against master `36762b5` and the `JOSS` branch.

Raised while reviewing for JOSS, openjournals/joss-reviews#11163.
