### 2026-08-28 graph-algorithms/edge-addition-planarity-suite#316 @MODSetter
`compile-warnings.yml` cannot fail on warnings that need `-O2`, and is hiding two `chdir` warnings
The compiler warning job passes, but that is an artefact of how `CFLAGS` is handed to `configure`
rather than evidence that the tree is clean.

`.github/workflows/compile-warnings.yml`:

```yaml
    env:
      CC: ${{ matrix.compiler }}
      CFLAGS: -Werror
    ...
          ./configure --enable-compile-warnings CC="$CC" CFLAGS="$CFLAGS"
```

Assigning `CFLAGS` on the `configure` line replaces autoconf's default `-g -O2` rather than adding to
it, so the job builds unoptimized. On Ubuntu that matters: gcc auto-defines `_FORTIFY_SOURCE` when
optimizing, fortification is what makes glibc attach `warn_unused_result` to `chdir`, and without
`-O2` none of that happens. So the class of warning the job exists to escalate never gets emitted.

Four configurations, same tree, Ubuntu 24.04 with gcc 13.3.0:

| `./configure` arguments | `make` exit | diagnostics |
|---|---|---|
| none, which is what a user runs | 0 | 2 warnings |
| `--enable-compile-warnings CC=gcc CFLAGS=-Werror`, the job verbatim | 0 | none |
| `--enable-compile-warnings CC=gcc CFLAGS="-O2 -Werror"` | 2, fails | 2 errors |
| `--enable-compile-warnings CC=gcc` | 0 | 2 warnings |

## The two warnings it is hiding

A plain `./configure && make` on Ubuntu today:

```
c/planarityApp/planarityCommandLine.c:218:17: warning: ignoring return value of 'chdir' declared with attribute 'warn_unused_result' [-Wunused-result]
c/planarityApp/planarityCommandLine.c:263:5: warning: ignoring return value of 'chdir' declared with attribute 'warn_unused_result' [-Wunused-result]
```

Both are the `chdir(origDir)` calls that restore the working directory. Line 218 is on the path where
the samples directory could not be entered, and line 263 runs after the test loop. If either fails the
process carries on in the wrong directory, which is a real, if unlikely, failure mode rather than
noise.

## Suggested fix

Two parts, either of which can go first:

1. Give the job the optimization level a user gets, `CFLAGS: "-O2 -Werror"`, or append `-Werror`
   through `AM_CFLAGS` so autoconf's `-g -O2` survives. Note this will fail the build until part 2
   lands.
2. Handle the two return values, either by checking them or by making the discard explicit:

```c
    if (chdir(origDir) != 0)
        gp_Message("WARNING: Unable to restore the original working directory.");
```

## Two caveats before you act on this

It is distro dependent, so it will not reproduce everywhere:

| Platform | compiler | does `-O2` define `_FORTIFY_SOURCE`? | `chdir` warning? |
|---|---|---|---|
| Ubuntu 24.04 x86_64 | gcc 13.3.0 | yes, Ubuntu patches gcc to default-enable it | yes |
| Ubuntu 24.04 aarch64 | gcc 13.3.0 | yes | yes, same two lines |
| Debian 12 | gcc 12.2.0 | no | no, but yes with explicit `-D_FORTIFY_SOURCE=2` |

It also cannot appear on macOS at all, since `warn_unused_result` on `chdir` is a glibc fortification
artefact. That is worth stating because a clean build on macOS or Debian is not evidence against this.

The job runs on `ubuntu-latest`, which is the one platform in the matrix where the warning does fire,
so it is pointed at the right target and then disarmed by its own `CFLAGS` assignment.

## The MSVC job has the same shape of problem

While checking the above I built with the `msvc` job's own command line from `dev-platforms.yml`. That
job compiles with `/Wall` and a suppression list but never passes `/WX`, so warnings cannot fail it
either. On the reviewed tree, MSVC 19.44 emits 14 diagnostics that the job reports as green:

```
c4127 conditional expression is constant
  c/planarityApp/planarityCommandLine.c(1671)
  c/graphLib/graph.c(2182), c/graphLib/graph.c(2212)
  c/graphLib/io/graphIO.c(305), c/graphLib/io/graphIO.c(329)

c4267 conversion from 'size_t' to 'int'/'long', possible loss of data
  c/planarityApp/planarityTestAllGraphs.c(265)
  c/planarityApp/planarityUtils.c(301), c/planarityApp/planarityUtils.c(971)
  c/graphLib/io/g6-read-iterator.c(471), (472), (562)
  c/graphLib/io/strbuf.c(187)
  c/graphLib/io/strOrFile.c(595), c/graphLib/io/strOrFile.c(720)
```

The nine `C4267` truncations only appear in the 64-bit configurations, where `size_t` is wider than
`int`; the `x86` build emits just the five `C4127`. All of them sit in string and file-length handling,
so reaching one needs an input over `INT_MAX`, and none is likely to be a live bug. They matter here
only as evidence that this second gate is also configured so a warning cannot fail it, which makes the
paper's "compiler-warning-sensitive automated tests" weaker than it reads.

Adding `/WX` would need those 14 addressed or suppressed first, so it is probably separate follow-up
work from the `chdir` fix above.

## Reproducing

```bash
git clone https://github.com/graph-algorithms/edge-addition-planarity-suite.git
cd edge-addition-planarity-suite
autoreconf -fi

./configure                                                   && make 2>&1 | grep warning:
make distclean
./configure --enable-compile-warnings CC=gcc CFLAGS=-Werror    && make 2>&1 | grep -c 'warning:\|error:'
make distclean
./configure --enable-compile-warnings CC=gcc CFLAGS="-O2 -Werror" ; make 2>&1 | grep error:
```

For the MSVC half, from a Developer Command Prompt, using the `dev-platforms.yml` command verbatim:

```
cl /O2 /nologo /Wall /wd4464 /wd4710 /wd4711 /wd4774 /wd4996 /wd5045 /wd4068 /wd4100 /wd4820 ^
 /external:anglebrackets /external:W0 /Fe./planarity.exe ./c/planarityApp/**.c ./c/graphLib/**.c ^
 ./c/graphLib/extensionSystem/**.c ./c/graphLib/homeomorphSearch/**.c ./c/graphLib/io/**.c ^
 ./c/graphLib/lowLevelUtils/**.c ./c/graphLib/planarityRelated/**.c
```

Verified against master `36762b5` and the `JOSS` branch at `ecfb06e`, both from fresh clones. The
`-O2 -Werror` configuration still fails on the same two `chdir` lines on current master.

Raised while reviewing for JOSS, openjournals/joss-reviews#11163.
