### 2026-08-28 graph-algorithms/edge-addition-planarity-suite#317 @MODSetter
Dev Setup wiki: the Debian dependency line does not install the autotools
The [Dev Setup wiki page](https://github.com/graph-algorithms/edge-addition-planarity-suite/wiki/2.-Dev-Setup)
gives this for Debian:

```
sudo apt update && sudo apt -y upgrade && sudo apt install -y build-essential gdb autotools-dev git-all
```

and describes it as installing "the GNU Autotools". It does not. `autotools-dev` ships two debhelper
scripts plus the `config.guess` and `config.sub` man pages. `autoreconf` comes from `autoconf`, and the
build also needs `automake` and `libtool`.

The result is that a reader following the documentation literally cannot get past the first build step
in the README. On a clean `debian:12` container with exactly that apt line:

```
=== tools the README's step 3.i needs ===
  gcc          /usr/bin/gcc
  make         /usr/bin/make
  git          /usr/bin/git
  autoreconf   NOT FOUND
  autoconf     NOT FOUND
  automake     NOT FOUND
  libtool      NOT FOUND
  aclocal      NOT FOUND

=== what autotools-dev actually ships in /usr/bin ===
  /usr/bin/dh_autotools-dev_restoreconfig
  /usr/bin/dh_autotools-dev_updateconfig

=== README step 3.i on a fresh clone ===
  autoreconf: command not found
  autoreconf exit: 127
```

## Suggested fix

```
sudo apt update && sudo apt -y upgrade && sudo apt install -y \
  build-essential gdb autoconf automake libtool git-all
```

With those three added, the same container gets through the whole documented sequence: `autoreconf -fi`,
`./configure`, `make` and `make check` all exit 0, with `PASS: test-samples.sh`.

## One suggestion beyond the fix

Consider putting a short Dependencies block in `README.md` as well. Right now the only dependency list
lives in the wiki, which means it is not versioned with the source and is not in the `make dist`
tarball, so someone building from a release tarball has nothing to go on.

## Reproducing

```bash
docker run --rm -it debian:12 bash
apt update && apt -y upgrade && apt install -y build-essential gdb autotools-dev git-all
git clone https://github.com/graph-algorithms/edge-addition-planarity-suite.git
cd edge-addition-planarity-suite && autoreconf -fi ; echo "exit: $?"
```

Checked on `debian:12` (the platform the wiki targets) and on Ubuntu 24.04. Wiki page re-read on
2026-08-28.

Raised while reviewing for JOSS, openjournals/joss-reviews#11163.
