### 2026-08-28 #11163 @MODSetter
@danielskatz @john-boyer-phd sorry again for the delay. My review is done and I recommend acceptance.

I worked mostly in containers on Ubuntu 24.04 with gcc 13.3.0, deliberately picking a different stack
from @Deln0r's macOS and clang pass so the two reviews do not overlap. What I checked directly:

- Built from a clean clone following the README literally, then `make check`, `make install`,
  `make dist` and `make distcheck`, all exit 0. Also `pip install planarity`, which builds the PyPI
  wrapper from its sdist through Cython; K5, K3,3 and Petersen come back non-planar, a balanced tree
  and a 5x5 grid planar, and `kuratowski_edges(K5)` returns all ten edges. One documentation snag,
  noted below.
- Reproduced the published reference tables for orders 8 through 11 with nauty `geng` and the in-repo
  orchestrators, and diffed them against the committed tables. Order 11 came out at 1,018,997,864
  graphs, 18,681,008 planar, 1,000,316,856 not, matching
  `TestSupport/tables/11/n11.mALL.p.out.txt` exactly, and the `-d`, `-o` and `-2` tables match at
  order 11 too. As a check that does not depend on your tables at all, the planar counts for orders 5
  to 11 agree with OEIS [A005470](https://oeis.org/A005470) and the graph totals with
  [A000088](https://oeis.org/A000088).
- Tested Windows natively, replicating the `msvc` job from `dev-platforms.yml` with MSVC 19.44 on
  Windows 11, x64 and x86. Both build and both pass `planarity.exe -test ./c/samples/`. I did this
  because neither of us had touched Windows and you ship four MSVC and five MSYS2 configurations.
- Replicated your `aarch64` job under emulation: `autoreconf -if && ./configure && make -j4 &&
  ./test-samples.sh` all clean, plus `make check`.
- Confirmed the paper's performance claim: `-rm 10000000` produces exactly 29,999,994 edges, embedded
  planar in 48 s with 3,968 MB peak memory.
- Checked obstruction isolation and planar drawing by hand on K5 and K3,3 rather than only trusting
  the batch counts, and drove the interactive menu mode through stdin as well as the command line.
- Compiled and linked a small third-party C program against the installed `libplanarity` and called
  `gp_Embed()` from it, with the include and link flags resolved only through
  `pkg-config --cflags --libs libplanarity`, to confirm the library is usable outside the bundled app.

I have ticked every item except **Installation instructions**, which I am holding on a one-line
documentation fix.

Rather than putting the detail here, I filed three issues on the repo:

- graph-algorithms/edge-addition-planarity-suite#315: signed integer overflow in
  `gp_CreateRandomGraphEx()`. `planarity -rm 46342` fails outright with exit 255 and no output file,
  while `-rm 46341` works and `-rm 10000000` works. Same boundary on Linux x86_64, Linux aarch64 and
  MSVC x64 and x86, so it is not platform specific. This is the one I would want fixed before the
  archive is cut.
- graph-algorithms/edge-addition-planarity-suite#316: neither warning gate can fail on a warning.
  `compile-warnings.yml` assigns `CFLAGS` on the `configure` line, which replaces autoconf's default
  `-g -O2`, so warnings that need `-O2` are never emitted; it is currently hiding two real `chdir`
  warnings that a default user build shows. The `msvc` job compiles `/Wall` but never passes `/WX`,
  so its 14 diagnostics are reported green too.
- graph-algorithms/edge-addition-planarity-suite#317: the Dev Setup wiki's Debian apt line is missing
  `autoconf`, `automake` and `libtool` while describing itself as installing the GNU Autotools, so a
  reader following it cannot run `autoreconf`. This is the only reason for my hold on Installation
  instructions.

None of the three blocks acceptance in my view. The first is a real bug, but it sits in a test data
generator rather than in the planarity algorithms, and no result in the paper depends on it.

Two comments on the paper, neither of which I filed as an issue. The first is the amended AI usage
disclosure, which reads well, and I have ticked that item. The absolute sentence is gone, "not aware
of any substantive use" is accurate against the history, and the VS Code explanation for the Copilot
trailers matches what those two commits actually contain. But the new text also asserts that the PR
review process and the "sanitizer and compiler-warning-sensitive automated tests" are what protect the
projects and their consumers from low quality contributions, and the two issues above are
counterexamples to that specific sentence. The overflow arrived through
[PR #279](https://github.com/graph-algorithms/edge-addition-planarity-suite/pull/279), which was
approved and merged from the head branch `codex/eaps-random-graph-ex-235` and is still in the tree a
month later. The sanitiser job only runs over `c/samples`, so it never sees a graph larger than 8
vertices and could not have caught it, and as of today neither warning gate can fail on a warning.
Either soften that claim or fix the two issues so it holds. I would rather see the second.

The second is smaller. The paper points readers at the
[`edge-addition-planarity-suite-testing` repo](https://github.com/graph-algorithms/edge-addition-planarity-suite-testing)
for the order-11 validation code, but that repo is a superseded copy. The tooling I actually used, and
which the committed tables are generated with, is in-tree at `TestSupport/planaritytesting/`; it has
six modules the standalone repo lacks, including the `testAllGraphs` orchestrator. Worth saying which
one is authoritative.

@danielskatz one note for the archive, and a correction to something I nearly got wrong. The reviewed
sources are `master`'s ancestor `0febfb2` ("Issue301 joss review (#302)", 2026-08-19), since
`git diff 0febfb2 origin/JOSS -- c/` is empty and the `JOSS` branch adds only the paper, the PDF
workflow and a `.gitignore` line on top of it. No tag points at that state. The existing
`Version_5.0.0.0` tag is from 2026-06-17 and sits 166 commits behind the `JOSS` branch with
substantial `c/` differences, so it is not the reviewed tree and should not be the archive target.
`master` has meanwhile moved 12 commits and 662 lines of `c/` ahead, including two features neither of
us looked at, the GraphML writer in
[#308](https://github.com/graph-algorithms/edge-addition-planarity-suite/pull/308) and the directed
DFS mode in [#309](https://github.com/graph-algorithms/edge-addition-planarity-suite/pull/309). So
@john-boyer-phd will need to cut a new tag. I have no objection to it including the 12 newer commits
if that is simpler, as long as it is stated, and I did at least confirm current `master` builds clean
and passes `make check` and the full sample suite.

For the record, what I did not cover, so the gaps are visible: no macOS, which is @Deln0r's half; no
MSYS2, Cygwin or Alpine/musl, though you have CI for all three; I did not re-benchmark LEDA or PIGALE,
so the comparison in the paper still rests on the 2003 publication, and LEDA being proprietary I do
not think a reviewer can close that one; and the two new features on `master` I smoke-tested rather
than read.

On AI assistance, as the checklist asks. I used an AI assistant for non-substantive help with this
review: the writing, and running and analysing the checks at my direction. The evaluation is mine, and
every judgement, every technical claim and the recommendation to accept was made and verified by me.
