### 2026-08-29 graph-algorithms/edge-addition-planarity-suite#315 @Deln0r
Different result on macOS, worth knowing before this gets judged by whether it fails.

Apple clang 21, arm64, master 36762b5, plain `-O2`: `-rm 46342` exits 0 and writes a correct 46,343-line file. So do 65536 and 1000000, which your table has failing under gcc and MSVC.

The UB is still there. UBSan flags the same line:

`c/graphLib/graph.c:1333:22: signed integer overflow: 46342 * 46341`

The difference is codegen. clang emits `mul` then `lsr`, a logical shift, since overflow being UB lets it assume a non-negative product. The wrapped pattern shifts back to the right half, so the clamp never fires. gcc's arithmetic shift keeps the sign and returns NOTOK.

Same bug, silent on the platform your `dev-platforms.yml` macOS job runs.
