### 2026-08-29 graph-algorithms/edge-addition-planarity-suite#316 @MODSetter
> Hi Rohan ([@MODSetter](https://github.com/MODSetter) ), This is a wonderful find, thank you for helping us to make this project better (regardless of whether you agree to do the following).
>
> In an August 18 post in [openjournals/joss-reviews#11163](https://github.com/openjournals/joss-reviews/issues/11163), Daniel said: "... reviewers are encouraged to submit issues and pull requests on the software repository." You've done such a good job identifying exactly what needs to be fixed for the `chdir()` problem, such that it would certainly be desirable to receive a PR from you so that you can be on the contributors list for making this contribution. If you choose to do a PR, please make sure your git configuration has the proper `user.name` and `user.email` because we have observed GitHub doesn't add to the contributor list unless those are set. **Please could you drop a comment here to let me know if you would like to do a PR or would prefer not?** In that latter case, no problem, I am grateful for this issue, and we will definitely fix this issue based on your contribution of this issue.
>
> _Edit_: Of the two proposed approaches (`CFLAGS -O2 -Werror` versus adding to `AM_FLAGS`), the latter approach seems preferable if it means that `-Werror` is added to `AX_APPEND_COMPILE_FLAGS` in `ax_compiler_flags_cflags.m4` because `-Werror` is one of the missing flags that I've mentioned in the second checklist item in [#310](https://github.com/graph-algorithms/edge-addition-planarity-suite/issues/310). We use more stringency in the VS Code tasks because those are what we happen to be using when reviewing PRs, but that additional stringency that's been added over time needs to be reflected in the makefile too, but that is the point of [#310](https://github.com/graph-algorithms/edge-addition-planarity-suite/issues/310), and just getting `Werror` in there now would be a positive step before [#310](https://github.com/graph-algorithms/edge-addition-planarity-suite/issues/310) is done.
>
> Finally, I would say it is reasonable to close this issue with a PR that addresses the `chdir` issue in the title and the first half of the main post because it would be reasonable to move the MSVC part to a separate issue. It would be important to file a separate issue only because I should probably take that one and make the 14 little decisions that will be needed to close each of the MSVC warnings. That sound OK to you, Rohan?

@john-boyer-phd I would love to raise a PR. I am always open to contributing to open source. Please give me a day or two. I will raise PRs for all my listed issues.
