### 2026-08-30 graph-algorithms/edge-addition-planarity-suite#315 @john-boyer-phd
This is a great find, thank you!

The lines in `gp_DynamicInsertEdge()` that contain `(N * (N - 1)) >> 1` can just be deleted (along with the preceding comment). It was originally written to put a cap on the number of edges at the number needed for a simple undirected graph with all edge connections between N vertices (a clique), but once we moved to supporting multiple parallel edges, I changed it to just making a pitstop at that number. But the pitstop serves no purpose, and now that it's clearly causing a problem because of integer overflow, it really needs to be deleted.

The suggested fix looks great for  `gp_CreateRandomGraphEx()`.

For `gp_GetRandomNumber()`, moving the `0x7FFFFFF` mask upward seems insufficient, so changing to unsigned for the folding looks to be the best approach. So, `N` could change to `unsigned` and then would be assigned `(unsigned) rand()` on the first line. The line that applies the`0x7FFFFFF` mask could be removed from before the modulus line, and the `0x7FFFFFF` mask could instead be applied in the `return` line, e.g., `return (int) ((N & 0x7FFFFFF) + NMin);`.
