### 2026-08-30 #11163 @john-boyer-phd
@MODSetter Regarding this: " But the new text also asserts that the PR review process and the "sanitizer and compiler-warning-sensitive automated tests" are what protect the projects and their consumers from low quality contributions, and the two issues above are counterexamples to that specific sentence. The overflow arrived through [PR #279](https://github.com/graph-algorithms/edge-addition-planarity-suite/pull/279), which was approved and merged from the head branch codex/eaps-random-graph-ex-235 and is still in the tree a month later. The sanitiser job only runs over c/samples, so it never sees a graph larger than 8 vertices and could not have caught it, and as of today neither warning gate can fail on a warning. Either soften that claim or fix the two issues so it holds. I would rather see the second. "

The high order bit is that we will be fixing the issues shortly, where "we" is actually including you, with thanks for your offering to do the PRs. I should clarify, though, that the overflow issue that you found in `gp_RandomGraphEx()` is not the fault of the new contributor nor of any AI agent. It is entirely my fault because it is the same calculation that I wrote years ago in `gp_DynamicAddEdge()` (see graph.c versions prior to August 22, when it was moved to `gp_DynamicInsertEdge()`). The new contributor reused the logic from that test in `gp_RandomGraphEx()`, and then I didn't flag it as a problem because I had done it that way for a long time, which was my mistake. Despite the fact that mistakes can get through a PR review, consumers are still well-protected by the processes, to the maximum extent that we are able. The processes are analogous to error correcting codes-- not perfect, but quite effective, especially compared to not having them.
