### 2026-08-31 graph-algorithms/edge-addition-planarity-suite#319 @Deln0r
gp_Read rejects a valid graph on unsigned-char targets when an extension is attached
## Describe the bug

`sf_getc()` and `sf_ungetc()` are declared to return `char` (`c/graphLib/io/strOrFile.h:42` and `:50`), and `sf_getc()` returns `EOF` as a `char` (`strOrFile.c:284`). Plain `char` is unsigned on AArch64, ARM, PowerPC and s390x, so on those targets `EOF` becomes 255 and every `== EOF` test against the result is decided by the type, not by the data.

At true end of input in `_ReadGraph()` (`graphIO.c:559`) the branch is therefore always taken, `sf_ungetc()` does not recognise 255 as `EOF` and pushes it back, and `sf_fgets()` hands the caller a one-byte string `{0xff, '\0'}` as extra data.

With the default post-processor this is invisible: extra data is ignored and the read succeeds, which is why the test suite passes on AArch64 and adjacency-list output is byte-identical to x86-64. Attach an extension first and the fabricated byte is delivered to its post-processor, which rejects it.

## Reproduction steps

`one.txt`, a valid one-vertex graph:

```text
N=1
1: -1
```

```c
graphP g = gp_New();
gp_ExtendWith_DrawPlanar(g);   /* supported before gp_Read; see graphDrawPlanar_Extensions.c:152 */
int r = gp_Read(g, "one.txt"); /* OK = 1 */
```

## Expected behavior

`gp_Read()` returns `OK` on every platform. The file is valid and carries no extra data.

## Terminal logs

Same source, same commit, same compiler, two container architectures:

```text
x86_64  (signed char)    no extension: 1    DrawPlanar: 1
aarch64 (unsigned char)  no extension: 1    DrawPlanar: 0
```

Without the extension both agree, which isolates the post-processor path rather than the graph read.

## Environment

- OS: Ubuntu 24.04 containers, `linux/amd64` and `linux/arm64`
- Built from source at `36762b5`, `autoreconf -if && ./configure && make`
- gcc 13.3.0 in both containers

## Additional context

`compile-warnings.yml` cannot see this. It runs `ubuntu-latest` only, where `char` is signed and the comparison is in range. The AArch64 leg of `dev-platforms.yml` does run on the affected target, but calls plain `./configure` without `--enable-compile-warnings`, and its pull-request trigger is commented out. Building x86-64 with `-funsigned-char` reproduces the diagnostic at `graphIO.c:560` and would give architecture-independent coverage in the existing gate.

I found this while testing #318 and checked it against unpatched `master`, so it is not caused by that change. Related: #316, which reports the same gate failing to fail.

The usual fix is `int` in both wrappers, keeping the value as `int` through the `EOF` check and narrowing only after it is known not to be `EOF`. I am happy to prepare that if you want it, though it touches every caller and is a larger change than the warnings work.
