### 2026-08-31 graph-algorithms/edge-addition-planarity-suite#319 @john-boyer-phd
Another great find, thank you Ian (@Deln0r ). The choice of char for `sf_getc()` and `sf_ungetc()` was my own suboptimal choice, because we currently have no supported file formats that should contain 0xFF (and I didn't know char could be unsigned!).

It's a larger but better "surgery" to flip over to `int`. If you're up for it, I understand it will take longer and am also up for detailed discussions and reviewing.

It so happens that the larger surgery *could* be deferred to a future issue with a simpler and good-if-not-perfect fix of always casting every occurrence of EOF to be (char) EOF. I'd be up for this approach also, because it would enable you to make a lower-risk fix in the code and then pull something into compiler-warnings.yml, either AArch64 or x86-64 with `-funsignedchar`. The only thing that would be lost by this is truncation of files containing 0xFF. It's not legal in any of the current formats, nor in UTF-8 encoding of the upcoming GraphML reader. It's true that ISO-8859-1 encoding of XML does allow 0xFF but it is a really obscure character, and we have an easy workaround to suggest for users (use UTF-8) until the surgery happens. It's a crawl-walk-run approach, since going from crawl to walk is safer and faster than going from crawl to run in one go.

Please let me know which initial approach you'd prefer, and thank you for offering to do help make this better.

Regardless of the initial approach you'd rather take for `sf_getc()` and `sf_ungetc()`, there is the matter of updating `compiler-warnings.yml`. This seems like the right job to update because it's the main one where I didn't comment out the automatic run on every PR (in July, `dev-platforms.yml` proved to be too expensive to run on every single PR, so instead I do a set of PRs then run it on master for the batch). Either of AArch64 or x86-64 with `-funsignedchar` works for me, but I got the sense that you might slightly prefer x86-64 with `-funsignedchar`. For that case, are you able to inject `-funsignedchar` from the outside (in the job and not `AX_APPEND_COMPILE_FLAGS`) without overriding the `-O2` and other compiler and warning flags?
