### 2026-08-31 graph-algorithms/edge-addition-planarity-suite#317 @wbkboyer
Excellent catch, @MODSetter !!  Upon performing the entire process to set up a Debian 12 VM and a Debian 13 VM using VirtualBox on a Windows 11 machine, I have verified that installing `autotools-dev` indeed doesn't install any of `autoconf`, `automake`, nor `libtoolize`. I rolled back my VM and ran the command you specified:

```
sudo apt update && sudo apt -y upgrade && sudo apt install -y \
  build-essential gdb autoconf automake libtool git-all
```

And was able to confirm the ability to build the release tarball + run `test-samples.sh` 🎉 Therefore, I've updated [2. Dev Setup](https://github.com/graph-algorithms/edge-addition-planarity-suite/wiki/2.-Dev-Setup) with this explicit dependency installation command.

---

As an aside, what I believe happened here is that because we run the `TestSupport` Python scripts, I ran the *optional* setup step (number 5) to install build dependencies for Python 3.12 on Debian 12; after running

```
sudo apt build-dep -y python3 && sudo apt install -y pkg-config
```

And without explicitly installing `autoconf`, `automake` and `libtool`, I now have `autoconf`, `automake` and `libtoolize`:

```
wbkboyer@debian12:~/edge-addition-planarity-suite$ ls /usr/bin/ | grep  autoconf
autoconf

wbkboyer@debian12:~/edge-addition-planarity-suite$ ls /usr/bin/ | grep  make
automake
automake-1.16
dh_makeshlibs
gmake
make
make-first-existing-target

wbkboyer@debian12:~/edge-addition-planarity-suite$ ls /usr/bin/ | grep libtool
libtoolize
```
