### 2026-09-01 graph-algorithms/edge-addition-planarity-suite#317 @john-boyer-phd
Closing this now because Wanda (@wbkboyer ) has fixed the wiki setup instruction and because I've now also added an entry to the `README.md` that implements the "one suggestion beyond the fix" in a way that will not go stale across releases.
