### 2026-09-01 #11163 @john-boyer-phd
@MODSetter Regarding this: "The second is smaller. The paper points readers at the
[edge-addition-planarity-suite-testing repo](https://github.com/graph-algorithms/edge-addition-planarity-suite-testing) for the order-11 validation code, but that repo is a superseded copy. The tooling I actually used, and which the committed tables are generated with, is in-tree at TestSupport/planaritytesting/; it has six modules the standalone repo lacks, including the testAllGraphs orchestrator. Worth saying which one is authoritative."

It turns out that that TestSupport/planaritytesting and EAPS-T are both equally authoritative for getting "test all graphs" results. Neither supersedes the other. The "test all graphs" logic was originally developed in EAPS at the `planarity.exe` application level, and TestSupport/planaritysupport orchestrates calls to a compiled `planarity.exe` to test all graphs in various Graph6 files. This was developed to completion before we had the Planarity 1.0.0 python package, which is also mentioned in the paper. Once we had that, though, we redeveloped the "test all graphs" logic in Python for a few reasons, one of which was to showcase the speed and reliability of the new Planarity 1.0.0 python package. This is what EAPS-T contains. We've tended so far to not check in the tables that EAPS-T generates because it is easy enough to regenerate them and check against those in TestSupport/planaritytesting. However, based on this comment, we have now amended the `README.md` in EAPS-T so that readers of the JOSS paper will know that they can get the checked-in table results from EAPS. So, the authoritative _tables_ are only committed in one place, which is the main EAPS repo.

TestSupport/planaritytesting does have a few other modules, but they are early-days experimentation modules not for test all graphs but for a more advanced testing methodology we have developed called Edge Deletion Analyzer. I call them experimentation modules because they helped us discover that it was Edge Deletion Analysis that we had to do. The EAPS-T also contains Edge Deletion Analyzer, in addition to Test All Graphs, but the little helper modules that led us to Edge Deletion Analysis are not needed in EAPS-T and are only retained in EAPS for ease of reference in a (draft) paper about it.
