### 2026-09-01 #11163 @john-boyer-phd
@MODSetter Regarding this: "I have ticked every item except Installation instructions, which I am holding on a one-line
documentation fix."

We have now fixed the issue 317 as recommended, including verification of why we didn't hit the problem before (we have tended to always do the optional `sudo apt` commands because we have to build the optional stuff too, and didn't notice that the required-only `sudo apt` was a bit underpowered). We also implemented something for the "beyond the fix" recommendation you made in that issue.

@danielskatz We agree that Rohan's filed bug issues 315 and 316 are easy to fix and should be fixed before we then cut a special tag for JOSS to use for archival purposes. I will update the project and lib version numbers on master before cutting that tag, so that the archived version will be clearly higher than the 5.0.0.0 release version originally submitted (because it will include fixes that have come out of this review).
