### 2026-09-01 graph-algorithms/edge-addition-planarity-suite#319 @Deln0r
I would rather do the int surgery. The cast approach leaves the 0xFF truncation in place and adds a mark at every call site, and since you are offering review, the larger fix is the better use of that. Plan: one MR flipping `sf_getc`/`sf_ungetc` and their callers to `int`, with a regression test that reads a file containing a literal 0xFF byte, then a small follow-up adding the CI leg once the tree is clean under it.

On your question: injecting `-funsigned-char` from the job env works and does not disturb the warning set, because `WARNING_CFLAGS` is assembled by configure separately from user CFLAGS. The `-O2` part is already lost today though: the existing `CFLAGS: -Werror` replaces autoconf's default `-g -O2`, which is what #316 reported. The tautological-comparison class fires at -O0 regardless, I verified that in a container, so the new leg can be `CFLAGS: -Werror -funsigned-char` for now. Restoring `-O2` in the gate is worth doing, but it turns the two chdir warnings from #316 into errors, so those two call sites need fixing first.
