### 2026-09-01 7eb3f78
Return int from sf_getc and sf_ungetc so EOF survives unsigned char (#319)
Plain char is unsigned on AArch64, ARM, PowerPC and s390x, where
(char)EOF == 255. sf_getc() therefore could not report EOF on those
platforms: _ReadGraph() fabricated a 0xFF extra-data byte at true end
of input and any attached extension's post-processor rejected a valid
file. Both wrappers now take and return int, following the stdio
getc()/ungetc() contract.

The conversion also fixes two latent defects in the other direction:
a literal 0xFF byte read from the string container sign-extended to
EOF on signed-char plat
[…]
