### 2026-09-01 graph-algorithms/edge-addition-planarity-suite#319 @john-boyer-phd
> I would rather do the int surgery. The cast approach leaves the 0xFF truncation in place and adds a mark at every call site, and since you are offering review, the larger fix is the better use of that. Plan: one MR flipping `sf_getc`/`sf_ungetc` and their callers to `int`, with a regression test that reads a file containing a literal 0xFF byte, then a small follow-up adding the CI leg once the tree is clean under it.
>

Thank you Ian (@Deln0r ), the “int” solution really is best.

> On your question: injecting `-funsigned-char` from the job env works and does not disturb the warning set, because `WARNING_CFLAGS` is assembled by configure separately from user CFLAGS. The `-O2` part is already lost today though: the existing `CFLAGS: -Werror` replaces autoconf's default `-g -O2`, which is what [#316](https://github.com/graph-algorithms/edge-addition-planarity-suite/issues/316) reported. The tautological-comparison class fires at -O0 regardless, I verified that in a container, so the new leg can be `CFLAGS: -Werror -funsigned-char` for now. Restoring `-O2` in the gate is worth doing, but it turns the two chdir warnings from [#316](https://github.com/graph-algorithms/edge-addition-planarity-suite/issues/316) into errors, so those two call sites need fixing first.

This sounds like the concern I actually have with -funsigned-char. Rohan will be fixing #316 , and when that happens, -Werror will go down into the m4 file, so there shouldn’t be a CFLAGS left for -funsigned-char to go into. And it seemed as though we have to get rid of CFLAGS in the job so that it doesn’t override -O2 and whatever else it is overriding. I know we don’t want -funsigned-char in the m4 either. So is there anyplace else it could go where it would wouldn’t affect all builds?
