### 2026-09-01 graph-algorithms/edge-addition-planarity-suite#319 @Deln0r
Yes, there is a clean spot. `Makefile.am` applies the warning set through `AM_CFLAGS = $(WARNING_CFLAGS)`, so the flag can ride the same plumbing behind its own configure switch, mirroring the existing one:

```
AC_ARG_ENABLE([unsigned-char],
    [AS_HELP_STRING([--enable-unsigned-char],
        [build with -funsigned-char, to test unsigned-plain-char ABIs on any host])])
AS_IF([test "x$enable_unsigned_char" = "xyes"],
    [AX_APPEND_COMPILE_FLAGS([-funsigned-char], [WARNING_CFLAGS])])
```

The CI leg then runs `./configure --enable-compile-warnings --enable-unsigned-char` with no CFLAGS at all: the autoconf default `-g -O2` survives, the m4 file stays free of the flag, and no other build is touched unless someone asks for it. The two other candidate spots each clobber something: `make AM_CFLAGS=...` replaces the warning set, and job CFLAGS is what we are retiring.

I can fold this switch into the CI follow-up after the int conversion lands.
