### 2026-09-01 #11163 @MODSetter
@danielskatz @john-boyer-phd my review is complete and I recommend acceptance.

[#317](https://github.com/graph-algorithms/edge-addition-planarity-suite/issues/317) is fixed, so I have ticked Installation instructions, my last open box. I re-read the Dev Setup
page instead of taking the fix on trust: the Debian line now installs `autoconf`, `automake` and
`libtool`, and the description no longer credits `autotools-dev` with installing the Autotools. Thanks
@wbkboyer for confirming it on Debian 12 and 13 VMs.

You are right about where the overflow came from, and I was wrong to frame it as arriving with [#279](https://github.com/graph-algorithms/edge-addition-planarity-suite/pull/279). My
own issue flagged the same expression in `gp_DynamicInsertEdge()` at `graph.c:1978-1979`, which dates
the pattern well before that PR, so I had the evidence and did not follow it. I withdraw that part, and
the error correcting code analogy is fair. I would still keep the coverage point, since the sanitiser
job never sees an N large enough to trip it and `-O2 -Werror` still fails on the two `chdir` lines.

"Superseded" was the wrong word for EAPS-T, and your history explains what I was looking at. The
amended README points readers at `TestSupport/tables/` for the committed tables, which is what I was
asking for.

Splitting the MSVC half of [#316](https://github.com/graph-algorithms/edge-addition-planarity-suite/issues/316) into its own issue works for me, since those 14 suppressions are your
calls, and `AM_CFLAGS` seems the better route now that [#318](https://github.com/graph-algorithms/edge-addition-planarity-suite/pull/318) has put `-Wunused-result` in the Autotools
set. I am starting on both PRs now. Separately, @Deln0r's [#319](https://github.com/graph-algorithms/edge-addition-planarity-suite/issues/319) looks like a real bug to me, though it
needs an extension attached before it surfaces, so it does not change my recommendation.

@danielskatz the archive is the one thing still waiting on you. The reviewed sources are `master`'s
ancestor `0febfb2`, nothing tags that state, and `master` has moved further ahead since I wrote that.
