### 2026-09-01 ec90721
Fix signed integer overflow in random graph generation at N >= 46342
