### 2026-09-01 3a68f44
Let compile-warnings.yml fail on warnings that need -O2, and handle the two chdir returns
