### 2026-09-01 efcb6e0
Name remaining C declaration parameters
