### 2026-09-01 f27e52d
Let compile-warnings.yml fail on warnings that need -O2, and handle the two chdir returns (#322)
Closes #316
