### 2026-09-01 27bf82c
Name remaining function declaration parameters (#312)
Closes #188

* Name graph function-table callback parameters

Match the callback parameter names to their implementations so editor hover hints expose the arguments used by extension hooks. The parameter types, order, and runtime behavior are unchanged.

* Name remaining C declaration parameters

---------

Co-authored-by: John M. Boyer <john.boyer.phd@gmail.com>
