### 2026-09-02 4e42fb2
Integrate read-error checks into planarity test runner
