### 2026-09-02 f3b93a3
Protect double edge capacity calculation from signed integer overflow
