### 2026-09-02 1e49719
Tweak test messages and run location
