### 2026-09-02 875b82b
Return int from sf_getc and sf_ungetc so EOF survives unsigned char (#320)
Closes #319

* Return int from sf_getc and sf_ungetc so EOF survives unsigned char (#319)

Plain char is unsigned on AArch64, ARM, PowerPC and s390x, where
(char)EOF == 255. sf_getc() therefore could not report EOF on those
platforms: _ReadGraph() fabricated a 0xFF extra-data byte at true end
of input and any attached extension's post-processor rejected a valid
file. Both wrappers now take and return int, following the stdio
getc()/ungetc() contract.

The conversion also fixes two latent defects in the other direction:

[…]
