### 2026-09-02 #11163 @john-boyer-phd
@MODSetter The fixes for #315 and #316 are now also in, including adding large graphs for the sanitizer to check. Also, Ian's (@Deln0r ) fix for #319 is also now on master too.

@danielskatz When you are back, could you let me know please what you think of this: It seems as though the best approach for archiving is for us to create a 5.1.0.0 release of planarity that has the things now on master, which includes some small features and bug fixes that have made it to master after the article was submitted as well as the improvements that have come from this review. The release would then have the tag that could be used for archiving.
