### 2026-09-03 a71c3b6
Add --enable-unsigned-char and an unsigned-char CI leg (#326)
Add --enable-unsigned-char and an unsigned-char CI leg (#319 follow-up)

The switch appends -funsigned-char to WARNING_CFLAGS via
AX_APPEND_COMPILE_FLAGS, riding the existing AM_CFLAGS plumbing, so no
CFLAGS assignment is needed in the workflow and the autoconf default
-g -O2 stays intact. Off by default; no build changes unless the switch
is given.

compile-warnings.yml gains one matrix leg, gcc-14 with the switch, so
char-signedness regressions of the issue-319 class surface on the
x86-64 runner on every pull request.
