### 2026-09-03 126ea37
Compile warnings should also test MSVC and should also test that the warning-free compiles do work
