### 2026-09-04 9916e3b
Fix MSVC complaint about assigning unsigned char to char in the 0xFF test
