### 2026-09-04 8fa7423
Fix MSVC warning about not testing for EOF during Graph6 file header validation
