### 2026-09-05 486f457
Add a sparse6 and incremental sparse6 read iterator (#38)
A new S6ReadIterator in c/graphLib/io/s6-read-iterator.[ch] reads nauty's
sparse6 (':' lines) and incremental sparse6 (';' lines) formats with the
same API shape as the graph6 read iterator: s6_NewReader(),
s6_InitReaderWithString(), s6_InitReaderWithFileName(), s6_ReadGraph(),
s6_EndReached() and s6_FreeReader(). gp_Read() and gp_ReadFromString()
dispatch a first line beginning with ':' or ';', or with the
">>sparse6<<" header, to it, so the transform command accepts sparse6
input directly.

The bit stream is decoded straight from the 
[…]
