+++ c/planarityApp/planarityCommandLine.c
@@ -43,6 +43,12 @@ int runDrawPlanarNonplanarWriteTest(void);
 int runReadWithExtensionAtEofTest(void);
 int runHighByteRoundTripTest(void);
 int runCapacityLimitTests(void);
+int runSparse6ReadTests(void);
+int runSparse6LockstepTest(char const *g6FileName, char const *s6FileName, int inputInMemFlag, int expectedNumGraphs);
+int runSparse6AcceptTest(char const *s6Str, char const *expectedG6Line);
+int runSparse6AcceptEdgesTest(char const *s6Str, int order, int const edges[][2], int numEdges);
+int runSparse6RejectTest(char const *s6Str);
+char *copySparse6TestString(char const *s6Str);
 int testDirectedDFS(void);
 int testPetersenDigraph(void);
 int testDigraphTranspose(void);
@@ -262,6 +268,8 @@ int runQuickRegressionTests(int argc, char *argv[])
     else if (runHighByteRoundTripTest() != OK)
         retVal = NOTOK;
     else if (runCapacityLimitTests() != OK)
+        retVal = NOTOK;
+    else if (runSparse6ReadTests() != OK)
         retVal = NOTOK;
     else if (runGraphMLTests() != OK)
         retVal = NOTOK;
@@ -747,6 +755,497 @@ int runCapacityLimitTests(void)
 
     if (Result == OK)
         gp_Message("Test succeeded.\n");
+
+    return Result;
+}
+
+/********************************************************************
+ runSparse6ReadTests()
+
+ Exercises the sparse6 and incremental sparse6 read iterator (EPIC #38).
+
+ N5-all.s6 and N5-all.inc.s6 were produced from N5-all.g6 by nauty's
+ copyg (-s and -i respectively), and nauty_example.s6 from
+ nauty_example.g6 in the same way, so the graph6 reader serves as
+ the oracle: every graph read through the sparse6 reader must encode
+ to the same graph6 line as the graph read from the .g6 file. The
+ string cases pin the corners of the encoding (the padding rules, an
+ incomplete final pair, CRLF, the four-byte order, the incremental
+ toggle) and the inputs the reader must reject.
+ ********************************************************************/
+
+int runSparse6ReadTests(void)
+{
+    int Result = OK;
+    unsigned origQuietMode = gp_GetQuietMode();
+    size_t i = 0;
+
+    // {sparse6 input, graph6 encoding of the last graph in it}
+    char const *acceptCases[][2] = {
+        // The example in the format specification
+        {":Fa@x^\n", "Fw??G"},
+        // No line terminator, and header with CRLF
+        {":Fa@x^", "Fw??G"},
+        {">>sparse6<<:Fa@x^\r\n", "Fw??G"},
+        // An incomplete final pair is padding
+        {":Dw\n", "D??"},
+        // Orders 2 and 4, where the padding rule for n = 2^k applies
+        {":A\n", "A?"},
+        {":An\n", "A_"},
+        {":C\n", "C?"},
+        {":Cw\n", "CC"},
+        {":CwN\n", "CE"},
+        {":CwI\n", "CF"},
+        {":Con\n", "CQ"},
+        {":Co`\n", "CU"},
+        {":Coa\n", "CT"},
+        {":Co`V\n", "CV"},
+        {":CoKN\n", "C]"},
+        {":CoKI\n", "C^"},
+        {":CcKI\n", "C~"},
+        // Incremental lines toggle edges relative to the previous graph
+        {":D\n;oN\n", "D?_"},
+        {":D\n;oN\n;oN\n", "D??"},
+        {":D\n;oN\n:D\n", "D??"},
+    };
+
+    // Order 63 uses the four-byte order encoding
[…]
