### 2026-09-05 8e46c4c
Affinity for symmetry when expressing the enforcement of a minimum (no functional effect)
