### 2026-09-05 f72c29e
Minor adjustments to runCapacityLimitTests()
When the API is correctly returning NOTOK, we suppress error messages. Currently, they only display in DEBUG mode, but they will eventually be emitted in release mode too (once about 1.000 error messages are added before each return NOTOK in a future issue).
