### 2026-09-05 eb2852c
Fix preexisting expressions to match capacity limit test.
