### 2026-09-05 eb9bb90
Validate vertex and edge capacities before int arithmetic (#327)
Closes #325 

* Validate vertex and edge capacities before int arithmetic (#325)

gp_EnsureVertexCapacity() now rejects a vertex count whose derived
capacity arithmetic cannot be represented in int, computing the
effective edge capacity as the greater of a pre-set edgeCapacity and
DEFAULT_EDGE_CAPACITY_FACTOR * N, per the design discussion in #325.
The stack bound (edgeCapacity << 2) + 2 strictly exceeds every other
derived quantity, so a single widened test on it shields the vertex
and edge storage bounds and the gp_UpperBoundEd
[…]
