### 2026-09-05 ed450ca
Addition of checks and error messages needed to propagate in all cases.
