### 2026-09-05 46327c0
Slight change of order in regression test calls.
