+++ c/planarityApp/planarityCommandLine.c
@@ -37,18 +37,18 @@ int runGraphTransformationTest(char const *command, char const *infileName, int 
 int runTestAllGraphsTest(char const *commandString, char const *infileName);
 int runHideRestoreTest(graphP theGraph);
 int runIdentifyContractTest(graphP theGraph);
-int runDigraphTests(void);
-int runGraphMLTests(void);
-int runDrawPlanarNonplanarWriteTest(void);
-int runReadWithExtensionAtEofTest(void);
-int runHighByteRoundTripTest(void);
-int runCapacityLimitTests(void);
 int testDirectedDFS(void);
 int testPetersenDigraph(void);
 int testDigraphTranspose(void);
+int runDigraphTests(void);
+int runDrawPlanarNonplanarWriteTest(void);
+int runReadErrorTests(void);
+int runReadWithExtensionAtEofTest(void);
+int runHighByteRoundTripTest(void);
+int runCapacityLimitTests(void);
 int runGraphMLWriteTest(char const *inputFileName, char const *expectedOutputFileName);
 int runBasicGraphMLWriteTest(void);
-int runReadErrorTests(void);
+int runGraphMLTests(void);
 
 /****************************************************************************
  Command Line Processor
@@ -258,6 +258,8 @@ int runQuickRegressionTests(int argc, char *argv[])
         retVal = NOTOK;
     else if (runDigraphTests() != OK)
         retVal = NOTOK;
+    else if (runReadErrorTests() != OK)
+        retVal = NOTOK;
     else if (runReadWithExtensionAtEofTest() != OK)
         retVal = NOTOK;
     else if (runHighByteRoundTripTest() != OK)
@@ -265,8 +267,6 @@ int runQuickRegressionTests(int argc, char *argv[])
     else if (runCapacityLimitTests() != OK)
         retVal = NOTOK;
     else if (runGraphMLTests() != OK)
-        retVal = NOTOK;
-    else if (runReadErrorTests() != OK)
         retVal = NOTOK;
 
     // All done.
