### 2026-09-05 d3d292e
Separate the test messages from the successor
