### 2026-09-05 8549d0b
Re-enable more tests for PR #313
