### 2026-09-05 87ae4cc
Fix for MSVC cl tasks run by GitHub Actions
