### 2026-09-06 8cb8e7f
Keep the edge storage dense after an incremental sparse6 deletion
gp_DeleteEdge() records a hole in the edge storage unless the pair it
removes is the last one, and DrawPlanar refuses to embed a graph whose
storage has holes (graphDrawPlanar.c checks edgeHoles before its edge
loops rather than testing each record). A graph read through the
incremental path could therefore not be drawn: with n8.mALL converted
to incremental sparse6, "planarity -t -d" failed on the ninth graph,
whose two edges sat in eighteen records with six holes, while the same
graphs in graph6 form drew fine.

After a deleti
[…]
