### 2026-09-06 57ca2a0
Finished fixing MSVC cl warnings for local VS Code builds.
