### 2026-09-06 68e547e
strbuf fixes for int overflow and MSVC cl (x86_amd64) warnings
