+++ c/graphLib/io/strbuf.c
@@ -22,7 +22,7 @@ strBufP sb_New(int capacity)
 {
     strBufP theStrBuf;
 
-    if (capacity < 0)
+    if (capacity < 0 || capacity == INT_MAX)
         return NULL;
 
     theStrBuf = (strBufP)malloc(sizeof(strBuf));
@@ -184,19 +184,32 @@ void sb_ReadSkipInteger(strBufP theStrBuf)
  ********************************************************************/
 int sb_ConcatString(strBufP theStrBuf, char const *s)
 {
-    int slen = s == NULL ? 0 : strlen(s);
+    // Switched to size_t to fix int overflow and also warnings on 64-bit builds
+    size_t strLen = s == NULL ? 0 : strlen(s);
 
-    if (slen == 0)
+    if (strLen == 0)
         return OK;
 
     if (theStrBuf == NULL || theStrBuf->buf == NULL)
         return NOTOK;
 
-    if (theStrBuf->size + slen > theStrBuf->capacity)
-    {
-        int newLen = theStrBuf->size + slen > 2 * theStrBuf->capacity ? theStrBuf->size + slen : 2 * theStrBuf->capacity;
-        char *newBuf = (char *)malloc((newLen + 1) * sizeof(char));
+    if (theStrBuf->size + strLen > INT_MAX)
+        return NOTOK;
 
+    if ((size_t)theStrBuf->size + strLen > (size_t)theStrBuf->capacity)
+    {
+        size_t newLen = 0, doubleCapacity = 0;
+        char *newBuf = NULL;
+
+        doubleCapacity = ((size_t)theStrBuf->capacity) << 1;
+        newLen = theStrBuf->size + strLen;
+        if (newLen < doubleCapacity && (doubleCapacity + 1) <= INT_MAX)
+            newLen = doubleCapacity;
+
+        if ((newLen + 1) > INT_MAX)
+            return NOTOK;
+
+        newBuf = (char *)malloc((newLen + 1) * sizeof(char));
         if (newBuf == NULL)
             return NOTOK;
 
@@ -207,7 +220,7 @@ int sb_ConcatString(strBufP theStrBuf, char const *s)
     }
 
     strcpy(theStrBuf->buf + theStrBuf->size, s);
-    theStrBuf->size += slen;
+    theStrBuf->size += (int)strLen;
 
     return OK;
 }
