### 2026-09-06 fb3e685
Fix two signed int overflow / MSVC cl (x86_amd64) warnings
