### 2026-09-06 0e7971d
MSVC cl x86_amd64 warning and int overflow fixes
