+++ c/graphLib/io/strOrFile.c
@@ -601,12 +601,15 @@ int sf_ungetc(int theChar, strOrFileP theStrOrFile)
 
 int sf_ungets(char *strToUnget, strOrFileP theStrOrFile)
 {
+    if (strToUnget == NULL || strlen(strToUnget) > INT_MAX)
+        return NOTOK;
+
     if (!sf_IsValidStrOrFile(theStrOrFile) ||
         theStrOrFile->containerType != INPUT_CONTAINER ||
         (int)strlen(strToUnget) > (sp_GetCapacity(theStrOrFile->ungetBuf) - sp_GetCurrentSize(theStrOrFile->ungetBuf)))
         return NOTOK;
 
-    for (int i = (strlen(strToUnget) - 1); i >= 0; i--)
+    for (int i = ((int)strlen(strToUnget) - 1); i >= 0; i--)
         // N.B. Convert through unsigned char so a 0xFF byte does not enter the
         // unget buffer as EOF on signed-char platforms
         sp_Push(theStrOrFile->ungetBuf, (unsigned char)strToUnget[i]);
@@ -722,9 +725,14 @@ int sf_fputs(char const *strToWrite, strOrFileP theStrOrFile)
     int outputLen = EOF;
 
     if (strToWrite == NULL ||
+        strlen(strToWrite) > INT_MAX ||
         !sf_IsValidStrOrFile(theStrOrFile) ||
         theStrOrFile->containerType != OUTPUT_CONTAINER)
+    {
+        gp_ErrorMessage("Internal error.");
+        sf_SetOutputErrorFlag(theStrOrFile);
         return EOF;
+    }
 
     // N.B. fputs() will fail and return EOF if pFile doesn't correspond
     // to an output stream
@@ -733,7 +741,7 @@ int sf_fputs(char const *strToWrite, strOrFileP theStrOrFile)
     else if (theStrOrFile->theStrBuf != NULL)
     {
         if (sb_ConcatString(theStrOrFile->theStrBuf, strToWrite) == OK)
-            outputLen = strlen(strToWrite);
+            outputLen = (int)strlen(strToWrite);
         else
             outputLen = EOF;
     }
@@ -758,7 +766,11 @@ int sf_WriteInteger(int intToWrite, strOrFileP theStrOrFile)
 
     if (!sf_IsValidStrOrFile(theStrOrFile) ||
         theStrOrFile->containerType != OUTPUT_CONTAINER)
+    {
+        gp_ErrorMessage("Internal error.");
+        sf_SetOutputErrorFlag(theStrOrFile);
         return NOTOK;
+    }
 
     if (theStrOrFile->pFile != NULL)
         result = fprintf(theStrOrFile->pFile, "%d", intToWrite) < 0 ? NOTOK : OK;
