### 2026-09-06 2c79d58
MSVC cl x64 and x86_amd64 warning fixes
