### 2026-09-06 b49094d
Version change for imminent release of 5.1.0.0 EAPS (5:0:1 graphLib/libPlanarity)
