### 2026-09-07 e090a13
Add CITATION.cff
