+++ CITATION.cff
@@ -1,0 +1,12 @@
+cff-version: 1.2.0
+message: If you use this software, please cite it as below.
+title: Edge Addition Planarity Suite and Generalized Graph Library
+type: software
+authors:
+- given-names: John M.
+  family-names: Boyer
+  orcid: https://orcid.org/0000-0002-4755-5535
+- given-names: Wanda B. K.
+  family-names: Boyer
+repository-code: https://github.com/graph-algorithms/edge-addition-planarity-suite
+license: BSD-3-Clause
