### 2026-09-07 831edd2
Propagate all string-or-file input failures
Distinguish clean EOF from read errors in integer parsing, make input failures sticky across the string-or-file helpers, and tag close failures according to container direction. Silence expected fault-injection diagnostics so the Valgrind log scan does not mistake them for sanitizer failures, and expand EOF/error-state coverage.
