+++ c/graphLib/io/strOrFile.c
@@ -298,20 +298,22 @@ int sf_getc(strOrFileP theStrOrFile)
     if (!sf_IsValidStrOrFile(theStrOrFile) ||
         theStrOrFile->containerType != INPUT_CONTAINER ||
         theStrOrFile->inputErrorFlag)
+    {
+        sf_SetInputErrorFlag(theStrOrFile);
         return EOF;
+    }
 
     if ((theStrOrFile->ungetBuf != NULL) && (sp_GetCurrentSize(theStrOrFile->ungetBuf) > 0))
     {
         int currChar = 0;
-        // Technically, this returns NOTOK on error, but the error is underflow, which is
-        // checked above and so cannot happen. Therefore, safe to use sp_Pop() here.
-
-#ifndef DEBUG
-        sp_Pop(theStrOrFile->ungetBuf, currChar);
+#ifndef SPEED_MACROS
+        if (sp__Pop(theStrOrFile->ungetBuf, &currChar) != OK)
+        {
+            sf_SetInputErrorFlag(theStrOrFile);
+            return EOF;
+        }
 #else
-        // In DEBUG, MSVC cl is misreading the printf() generated by sp_Pop(),
-        // so we cut to the underlying debug method to stop its warnings
-        sp__Pop(theStrOrFile->ungetBuf, &(currChar));
+        sp_Pop(theStrOrFile->ungetBuf, currChar);
 #endif
         theChar = currChar;
     }
@@ -320,7 +322,7 @@ int sf_getc(strOrFileP theStrOrFile)
         theChar = getc(theStrOrFile->pFile);
         if (ferror(theStrOrFile->pFile))
         {
-            theStrOrFile->inputErrorFlag = TRUE;
+            sf_SetInputErrorFlag(theStrOrFile);
             return EOF;
         }
     }
@@ -328,11 +330,14 @@ int sf_getc(strOrFileP theStrOrFile)
     {
         // N.B. Convert through unsigned char so that a literal 0xFF byte in
         // the string buffer does not sign-extend to EOF on signed-char platforms
-        theChar = sb_GetReadString(theStrOrFile->theStrBuf) != NULL
-                      ? (unsigned char)sb_GetReadString(theStrOrFile->theStrBuf)[0]
-                      : EOF;
-        if (theChar != EOF)
-            sb_ReadSkipChar(theStrOrFile->theStrBuf);
+        char *readString = sb_GetReadString(theStrOrFile->theStrBuf);
+        if (readString == NULL)
+        {
+            sf_SetInputErrorFlag(theStrOrFile);
+            return EOF;
+        }
+        theChar = (unsigned char)readString[0];
+        sb_ReadSkipChar(theStrOrFile->theStrBuf);
     }
 
     return theChar;
@@ -348,10 +353,16 @@ int sf_ReadSkipChar(strOrFileP theStrOrFile)
 {
     if (!sf_IsValidStrOrFile(theStrOrFile) ||
         theStrOrFile->containerType != INPUT_CONTAINER)
+    {
+        sf_SetInputErrorFlag(theStrOrFile);
         return NOTOK;
+    }
 
     if (sf_getc(theStrOrFile) == EOF)
+    {
+        sf_SetInputErrorFlag(theStrOrFile);
         return NOTOK;
+    }
 
     return OK;
 }
@@ -369,7 +380,10 @@ int sf_ReadSkipWhitespace(strOrFileP theStrOrFile)
 
     if (!sf_IsValidStrOrFile(theStrOrFile) ||
         theStrOrFile->containerType != INPUT_CONTAINER)
+    {
+        sf_SetInputErrorFlag(theStrOrFile);
         return NOTOK;
[…]
