+++ c/planarityApp/planarity-test/testReadErrors.c
@@ -28,6 +28,7 @@ typedef struct
     const char *data;
     size_t position;
     int failAtEnd;
+    int failOnClose;
 } TestInput;
 
 static int failures = 0;
@@ -70,16 +71,28 @@ static int readInput(void *cookie, char *buffer, int size)
     return 0;
 }
 
+static int closeInput(void *cookie)
+{
+    TestInput *input = (TestInput *)cookie;
+
+    if (input->failOnClose)
+    {
+        errno = EIO;
+        return EOF;
+    }
+    return 0;
+}
+
 static int newInput(TestInput *input, strOrFileP *pContainer)
 {
     strOrFileP container = sf_NewInputContainer("placeholder", NULL);
     FILE *stream = NULL;
 
 #ifdef HAVE_FOPENCOOKIE
-    cookie_io_functions_t functions = {readInput, NULL, NULL, NULL};
+    cookie_io_functions_t functions = {readInput, NULL, NULL, closeInput};
     stream = fopencookie(input, "r", functions);
 #else
-    stream = funopen(input, readInput, NULL, NULL, NULL);
+    stream = funopen(input, readInput, NULL, NULL, closeInput);
 #endif
     if (container == NULL || stream == NULL)
     {
@@ -100,7 +113,7 @@ static int testLineReads(void)
     char buffer[32] = {0};
     char pushed[] = "pushed";
     char retry[] = "retry";
-    TestInput input = {"line\n", 0, FALSE};
+    TestInput input = {"line\n", 0, FALSE, FALSE};
     strOrFileP container = NULL;
 
     if (newInput(&input, &container) != OK)
@@ -161,8 +174,17 @@ static int testLineReads(void)
 
 static int testCharacterReads(void)
 {
-    TestInput input = {"x", 0, TRUE};
+    TestInput input = {"x", 0, TRUE, FALSE};
     strOrFileP container = NULL;
+    const struct
+    {
+        const char *data;
+        int value;
+    } integers[] = {
+        {"123", 123},
+        {"-7", -7},
+        {"214748364", 214748364},
+        {"-214748364", -214748364}};
 
     if (newInput(&input, &container) != OK)
         return NOTOK;
@@ -173,16 +195,19 @@ static int testCharacterReads(void)
 
     for (int failAtEnd = FALSE; failAtEnd <= TRUE; failAtEnd++)
     {
-        int value = 77;
-        input.data = "123";
-        input.position = 0;
-        input.failAtEnd = failAtEnd;
-        if (newInput(&input, &container) != OK)
-            return NOTOK;
-        CHECK(sf_ReadInteger(&value, container) == (failAtEnd ? NOTOK : OK));
-        CHECK(value == (failAtEnd ? 77 : 123));
-        CHECK(container->inputErrorFlag == failAtEnd);
-        sf_Free(&container);
+        for (size_t index = 0; index < sizeof(integers) / sizeof(integers[0]); index++)
+        {
+            int value = 77;
+            input.data = integers[index].data;
+            input.position = 0;
+            input.failAtEnd = failAtEnd;
+            if (newInput(&input, &container) != OK)
+                return NOTOK;
+            CHECK(sf_ReadInteger(&value, container) == (failAtEnd ? NOTOK : OK));
+            CHECK(value == (failAtEnd ? 77 : integers[index].value));
+            CHECK(container->inputErrorFlag == failAtEnd);
+            sf_Free(&container);
+        }
 
         input.data = " \t";
         input.position = 0;
@@ -192,6 +217,70 @@ static int testCharacterReads(void)
[…]
