### 2026-09-08 d620908
Add unget buffer overflow check
