### 2026-09-11 ae62f67
minor changes in paper.md
