+++ JOSS/paper.md
@@ -45,7 +45,7 @@ Graphs are used to model a very wide array of real-world problems in which there
 are objects and relationships between the objects. In artificial intelligence,
 graphs are used to help with reasoning tasks, such as about the relationship
 pathways between persons of interest in law enforcement or between genes,
-tissues, diseases, and medications in bioinformatics research. In Physics,
+tissues, diseases, and medications in bioinformatics research. In physics,
 graphs and planarity are used to help compute material phase transitions,
 particle interactions, and electromagnetic duality. Similarly, in chemistry,
 graphs may be used to represent atoms and their valence bonds in molecules.
@@ -55,7 +55,7 @@ a graph may be used to represent the electrical components and wiring in a
 circuit, such as transistors and etchings on a silicon wafer. The graph of a
 circuit must be planar or planarized to eliminate short circuits.
 
-The open source software described in this paper provides a highly performant
+The open-source software described in this paper provides a highly-performant
 generalized graph library that also implements state-of-the-art algorithms for
 planarity-related problems. The generalized graph library and planarity-related
 algorithms are available to C/C++ and Python graph application developers.
@@ -64,7 +64,7 @@ algorithms are available to C/C++ and Python graph application developers.
 
 The `Edge Addition Planarity Suite` is an open source software package that was
 originally developed to implement a new planarity algorithm in
-[@BoyerMyrvold:EdgeAddition]. The Edge Addition Planarity Algorithm is currently
+@BoyerMyrvold:EdgeAddition. The Edge Addition Planarity Algorithm is currently
 regarded as one of two state-of-the-art planarity algorithms due to its
 conceptual simplicity as well as its speed of execution
 [@Wikipedia:PlanarityTesting]. The software package is available from the
@@ -73,8 +73,8 @@ repository](https://github.com/graph-algorithms/edge-addition-planarity-suite)
 under a 3-clause BSD license. 
 
 To enable development of the Edge Addition Planarity Algorithm, as well as the
-other planarity-related algorithms in [@Boyer:GD2005] and
-[@Boyer:SubgraphHomeomorphism], it was necessary to develop an extensible,
+other planarity-related algorithms in @Boyer:GD2005 and
+@Boyer:SubgraphHomeomorphism, it was necessary to develop an extensible,
 generalized graph library. This generalized graph library provides a
 high-performance solution for the graph algorithm representation and processing
 needs of a very wide array of graph applications. 
@@ -97,11 +97,11 @@ are not meant for high performance graph algorithms, so the Library for
 Efficient Data Types and Algorithms
 ([LEDA](https://leda.uni-trier.de/leda/guide/Index.html)) is significantly
 faster than they are because it is an industrial library designed for
-computational efficiency. Yet, in [@BoyerCortesePatrignaniDiBattista:GD2003],
+computational efficiency. Yet, in @BoyerCortesePatrignaniDiBattista:GD2003,
 all planarity algorithms implemented in LEDA were shown  to be several times
 slower than the Edge Addition Planarity Algorithm implementation. LEDA also has
 the disadvantage of being C++ only and has a proprietary license. The only graph
-library shown (in [@BoyerCortesePatrignaniDiBattista:GD2003]) to have similar
+library shown (in @BoyerCortesePatrignaniDiBattista:GD2003) to have similar
 performance to the Edge Addition Planarity Suite is the Public Implementation of
 a Graph Algorithm Library and Editor
 ([PIGALE](https://sourceforge.net/projects/pigale/)). However, PIGALE is also
[…]
