### 2026-09-13 a7e9c38
Hyphenation edit in JOSS paper markdown
