+++ CITATION.cff
@@ -1,12 +1,34 @@
-cff-version: 1.2.0
-message: If you use this software, please cite it as below.
-title: Edge Addition Planarity Suite and Generalized Graph Library
-type: software
+cff-version: "1.2.0"
 authors:
-- given-names: John M.
-  family-names: Boyer
-  orcid: https://orcid.org/0000-0002-4755-5535
-- given-names: Wanda B. K.
-  family-names: Boyer
-repository-code: https://github.com/graph-algorithms/edge-addition-planarity-suite
-license: BSD-3-Clause
+- family-names: Boyer
+  given-names: John M.
+  orcid: "https://orcid.org/0000-0002-4755-5535"
+- family-names: Boyer
+  given-names: Wanda B. K.
+contact:
+- family-names: Boyer
+  given-names: John M.
+  orcid: "https://orcid.org/0000-0002-4755-5535"
+doi: 10.5281/zenodo.22682932
+message: If you use this software, please cite our article in the
+  Journal of Open Source Software.
+preferred-citation:
+  authors:
+  - family-names: Boyer
+    given-names: John M.
+    orcid: "https://orcid.org/0000-0002-4755-5535"
+  - family-names: Boyer
+    given-names: Wanda B. K.
+  date-published: 2026-09-13
+  doi: 10.21105/joss.11163
+  issn: 2475-9066
+  issue: 125
+  journal: Journal of Open Source Software
+  publisher:
+    name: Open Journals
+  start: 11163
+  title: Edge Addition Planarity Suite and Generalized Graph Library
+  type: article
+  url: "https://joss.theoj.org/papers/10.21105/joss.11163"
+  volume: 11
+title: Edge Addition Planarity Suite and Generalized Graph Library
