### 2026-09-16 a186c90
test: cover close errors and truncated graph6 input
