### 2026-09-19 995cb2b
Test algorithms on sparse6 and incremental sparse6 input
planarity -t read .g6 input only. It now reads the format the input file
actually contains, so the sparse6 read iterator added for issue #38 is
usable from the command that consumes whole files of graphs.

- Move the first-line test that gp_Read() uses into s6_IsSparse6Input(),
  so the two callers cannot drift apart.
- Drive either read iterator through a small facade in
  planarityTestAllGraphs.c. The input is opened once: the first line is
  pushed back into the input container and the container is handed to
  the iterator that matche
[…]
