+++ c/graphLib/io/graphIO.c
@@ -25,6 +25,7 @@ See the LICENSE.TXT file for licensing information.
 extern int _g6_ReadGraphFromStrOrFile(graphP theGraph, strOrFileP *pG6InputContainer);
 extern int _g6_WriteGraphToStrOrFile(graphP theGraph, strOrFileP *pOutputContainer);
 extern int _s6_ReadGraphFromStrOrFile(graphP theGraph, strOrFileP *pS6InputContainer);
+extern int s6_IsSparse6Input(char const *const firstLine);
 extern int _WriteGraphMLGraph(graphP theGraph, strOrFileP outputContainer);
 
 /* Private functions (exported to system) */
@@ -541,8 +542,7 @@ int _ReadGraph(graphP theGraph, strOrFileP *pInputContainer)
         if (RetVal == OK)
             extraDataAllowed = TRUE;
     }
-    else if (lineBuff[0] == ':' || lineBuff[0] == ';' ||
-             strncmp(lineBuff, ">>sparse6<<", strlen(">>sparse6<<")) == 0)
+    else if (s6_IsSparse6Input(lineBuff))
     {
         // A ';' first line is an error that the sparse6 reader reports.
         // As for .g6 below, ownership of inputContainer passes to the
