+++ c/planarityApp/planarityTestAllGraphs.c
@@ -5,6 +5,7 @@ See the LICENSE.TXT file for licensing information.
 */
 
 #include "planarity.h"
+#include "../graphLib/io/strOrFile.h"
 
 typedef struct
 {
@@ -19,6 +20,30 @@ typedef testAllStats *testAllStatsP;
 
 int testAllGraphs(char command, char modifier, char const *const infileName, testAllStatsP stats);
 int outputTestAllGraphsResults(char command, char modifier, testAllStatsP stats, char const *const infileName, char *outfileName, char **pOutputStr);
+
+// A .g6 file is read by the graph6 read iterator and a .s6 or .inc.s6 file by
+// the sparse6 read iterator, so the loop below drives whichever one matches the
+// content of the input file through this small facade.
+typedef struct
+{
+    G6ReadIteratorP g6ReadIterator;
+    S6ReadIteratorP s6ReadIterator;
+} allGraphsReader;
+
+typedef allGraphsReader *allGraphsReaderP;
+
+// Package private initializers of the read iterators. The facade needs them
+// because the format of the input is decided from its first line, which has to
+// be read before the iterator exists. Re-opening the input by name instead
+// would read the first line twice, and an input that cannot be re-opened, such
+// as stdin or a pipe, would lose its first graph.
+extern int _g6_InitReaderWithStrOrFile(G6ReadIteratorP theG6ReadIterator, strOrFileP *pInputContainer);
+extern int _s6_InitReaderWithStrOrFile(S6ReadIteratorP theS6ReadIterator, strOrFileP *pInputContainer);
+
+static int allGraphsReaderInit(allGraphsReaderP theReader, graphP theGraph, char const *const infileName);
+static int allGraphsReaderReadGraph(allGraphsReaderP theReader);
+static int allGraphsReaderEndReached(allGraphsReaderP theReader);
+static void allGraphsReaderFree(allGraphsReaderP theReader);
 
 // #define TESTALLGRAPHS_MEMORY_TIMING_TEST
 
@@ -97,6 +122,117 @@ int TestAllGraphs(char const *const commandString, char const *const infileName,
     return Result;
 }
 
+/****************************************************************************
+ allGraphsReaderInit()
+
+ Opens the input, decides its format from its first line by the same test that
+ gp_Read() applies, and creates the matching read iterator.
+
+ The input is opened once: the first line is pushed back into the input
+ container and the container is then transferred to the iterator, so that
+ inputs which cannot be re-opened, such as stdin, are read from their first
+ byte exactly once. An input whose first line cannot be read is left to the
+ graph6 iterator, which reports the failure.
+
+ On failure the caller frees the facade with allGraphsReaderFree(), which
+ handles a partially created iterator.
+ ****************************************************************************/
+static int allGraphsReaderInit(allGraphsReaderP theReader, graphP theGraph, char const *const infileName)
+{
+    strOrFileP inputContainer = NULL;
+    char lineBuff[MAXLINE + 1];
+    int Result = OK;
+
+    if (theReader == NULL)
+        return NOTOK;
+
+    theReader->g6ReadIterator = NULL;
+    theReader->s6ReadIterator = NULL;
+
+    if ((inputContainer = sf_NewInputContainer(NULL, infileName)) == NULL)
+    {
+        gp_ErrorMessage("Unable to open \"%.*s\" for reading.",
+                        FILENAME_MAX, infileName);
+        return NOTOK;
+    }
+
+    memset(lineBuff, '\0', (MAXLINE + 1));
+
+    if (sf_fgets(lineBuff, MAXLINE, inputContainer) != NULL &&
+        sf_ungets(lineBuff, inputContainer) != OK)
+    {
[…]
