### 2026-09-19 ecb35b1
Keep sf_fgets() inside the caller's buffer when reading pushback
sf_fgets() read up to count characters from the pushback buffer and
wrote the null terminator after them, so a pushback buffer holding at
least count characters made it write str[count]. Its file path already
follows stdio and stores at most count - 1 characters, so the two paths
of the same function disagreed on the bound.

gp_Read() reaches the case: it reads the first line of the input and
pushes it back for the reader that the line selects, so a .g6 file whose
first line is longer than its order allows overflows the graph6 re
[…]
