### 2026-09-20 d3011e4
Improve readability and operation of sf_fgets()
