+++ c/graphLib/io/strOrFile.c
@@ -743,10 +743,11 @@ int sf_ungets(char *strToUnget, strOrFileP theStrOrFile)
 
 char *sf_fgets(char *str, int count, strOrFileP theStrOrFile)
 {
-    int charsToReadFromUngetBuf = 0;
-    int charsToReadFromStrOrFile = count;
+    int charsReadFromUngetBuf = 0;
+    int countForStrOrFile = count;
 
-    if (str == NULL || count < 0 ||
+    // Count must be positive, just like fgets()
+    if (str == NULL || count <= 0 ||
         !sf_IsValidStrOrFile(theStrOrFile) ||
         theStrOrFile->containerType != INPUT_CONTAINER ||
         theStrOrFile->inputErrorFlag)
@@ -755,6 +756,10 @@ char *sf_fgets(char *str, int count, strOrFileP theStrOrFile)
         return NULL;
     }
 
+    // Initialize str to an empty string
+    str[0] = '\0';
+
+    // Read characters from the unget buffer, if any
     if (theStrOrFile->ungetBuf != NULL)
     {
         int numCharsInUngetBuf = sp_GetCurrentSize(theStrOrFile->ungetBuf);
@@ -768,10 +773,10 @@ char *sf_fgets(char *str, int count, strOrFileP theStrOrFile)
             // stays inside a buffer of count bytes. Without this bound, a
             // pushback buffer holding at least count characters made the
             // loop below write str[count].
-            int maxCharsToRead = (count > 0) ? (count - 1) : 0;
+            int maxCharsToRead = count - 1;
+            int charsToReadFromUngetBuf = (maxCharsToRead > numCharsInUngetBuf) ? numCharsInUngetBuf : maxCharsToRead;
 
-            charsToReadFromUngetBuf = (maxCharsToRead > numCharsInUngetBuf) ? numCharsInUngetBuf : maxCharsToRead;
-            for (int i = 0; i < charsToReadFromUngetBuf; i++)
+            for (charsReadFromUngetBuf = 0; charsReadFromUngetBuf < charsToReadFromUngetBuf; charsReadFromUngetBuf++)
             {
                 currChar = sf_getc(theStrOrFile);
                 if (currChar == EOF)
@@ -779,8 +784,8 @@ char *sf_fgets(char *str, int count, strOrFileP theStrOrFile)
                     sf_SetInputErrorFlag(theStrOrFile);
                     return NULL;
                 }
-                str[i] = (char)currChar;
-                str[i + 1] = '\0';
+                str[charsReadFromUngetBuf] = (char)currChar;
+                str[charsReadFromUngetBuf + 1] = '\0';
                 // N.B. fgets() includes the \n in the string returned, and
                 // no further characters shall be read
                 if (currChar == '\n')
@@ -791,42 +796,67 @@ char *sf_fgets(char *str, int count, strOrFileP theStrOrFile)
             }
             // N.B. If we broke out of the loop early due to \n, do not read
             // any further characters from stream
-            charsToReadFromStrOrFile = (encounteredNewline) ? 0 : ((maxCharsToRead > numCharsInUngetBuf) ? (count - charsToReadFromUngetBuf) : 0);
+            if (encounteredNewline)
+                countForStrOrFile = 0;
+            // If the required number of characters have all been read from the
+            // unget buffer, then don't read any more from the stream.
+            else if (maxCharsToRead <= charsReadFromUngetBuf)
+                countForStrOrFile = 0;
+            // We need to read more characters from the stream, in which case
+            // we need a countForStrOrFile based on count, which is one greater
+            // than the number of chars needed, for the null terminator.
+            else
+                countForStrOrFile = count - charsReadFromUngetBuf;
         }
     }
 
-    if (charsToReadFromStrOrFile > 0)
+    if (countForStrOrFile > 0)
     {
[…]
