### 2026-09-21 3d34ce3
Promote the read iterator that chooses the format into graphLib
The facade that let planarity -t read graph6 and sparse6 lived in the
application. It belongs beside the two read iterators it drives, so
that gp_Read(), the application and the Python side of EAPS-T can all
decide the format the same way, and so that a later format is added in
one place.

- c/graphLib/io/gp-read-iterator.[ch]: GPReadIterator, with the API
  shape of the graph6 and sparse6 read iterators under the gp_ prefix.
  The type of the input is decided once, when the iterator is
  initialized, and stored in the iterator as
[…]
