+++ c/planarityApp/planarityCommandLine.c
@@ -38,6 +38,7 @@ int runTestAllGraphsTest(char const *commandString, char const *infileName, char
 int runHideRestoreTest(graphP theGraph);
 int runIdentifyContractTest(graphP theGraph);
 int runSparse6ReadTests(void);
+int runGeneralReadIteratorTests(void);
 int runSparse6LockstepTest(char const *g6FileName, char const *s6FileName, int inputInMemFlag, int expectedNumGraphs);
 int runSparse6AcceptTest(char const *s6Str, char const *expectedG6Line);
 int runSparse6AcceptEdgesTest(char const *s6Str, int order, int const edges[][2], int numEdges);
@@ -273,6 +274,8 @@ int runQuickRegressionTests(int argc, char *argv[])
     else if (runCapacityLimitTests() != OK)
         retVal = NOTOK;
     else if (runSparse6ReadTests() != OK)
+        retVal = NOTOK;
+    else if (runGeneralReadIteratorTests() != OK)
         retVal = NOTOK;
     else if (runGraphMLTests() != OK)
         retVal = NOTOK;
@@ -783,6 +786,269 @@ int runCapacityLimitTests(void)
  incomplete final pair, CRLF, the four-byte order, the incremental
  toggle) and the inputs the reader must reject.
  ********************************************************************/
+
+/****************************************************************************
+ runGeneralReadIteratorTests()
+
+ The general read iterator decides the format of an input from its first
+ line and then reads it with the read iterator of that format, so these
+ cases check that the decision is made correctly, that it is made once,
+ and that an input in none of the formats is refused rather than handed
+ to a reader that cannot make sense of it.
+ ****************************************************************************/
+
+int runGeneralReadIteratorTests(void)
+{
+    int Result = OK;
+    unsigned origQuietMode = gp_GetQuietMode();
+    size_t i = 0;
+
+    // {input string, graph6 encoding of the first graph it contains}
+    char const *acceptCases[][2] = {
+        // graph6, with and without its header
+        {"D?{\n", "D?{"},
+        {">>graph6<<D?{\n", "D?{"},
+        // sparse6 and incremental sparse6, the latter with the whole graph
+        // that its second line modifies
+        {":Fa@x^\n", "Fw??G"},
+        {":D\n;oN\n", "D??"},
+        {">>sparse6<<:Fa@x^\n", "Fw??G"},
+    };
+
+    // Inputs in none of the formats the iterator reads
+    char const *rejectCases[] = {
+        // digraph6, which neither reader accepts
+        "&AG\n",
+        // GraphML, whose first character is below the graph6 range
+        "<graphml>\n",
+    };
+
+    // graph6 marks its lines with nothing but the order, so any first byte in
+    // its range begins what could be a graph6 line: an adjacency list, which
+    // gp_Read() recognizes by its "N=" but a file of graphs cannot contain, is
+    // taken for graph6 and then fails on the content of the line.
+    char const *readFailCases[] = {
+        "N=3\n1: 2 3 0\n2: 1 0\n3: 1 0\n",
+    };
+
+    char const *g6Header = ">>graph6<<";
+    graphP theGraph = NULL;
+    GPReadIteratorP theReadIterator = NULL;
+    char *actualEncoding = NULL;
+
+    // {first line, expected s6_IsSparse6Input(), expected g6_IsGraph6Input()}
+    struct
+    {
+        char const *firstLine;
+        int isSparse6;
+        int isGraph6;
+    } predicateCases[] = {
+        {">>sparse6<<:Fa@x^", TRUE, FALSE},
[…]
