### 2026-09-21 fd76d9e
Refuse to allocate over an existing read iterator, and pin the refusals
gp_NewReader() overwrote a non-NULL iterator pointer, where the graph6
and sparse6 constructors refuse it, so a second call lost the first
iterator with its child, input container and buffers.

The tests now cover it: after a refused second initialization they check
that the iterator was not allocated over and that it still reads the
next graph, the predicate table gains the byte after the graph6 range
and one with the high bit set, and the file cases read every graph of
the sample to the end rather than only its first.
