### 2026-09-22 8049dbc
Reject an adjacency list at the graph6 predicate, and dispatch by switch
g6_IsGraph6Input() now requires the byte after the order byte to be in
the graph6 range as well, or to end the line. An adjacency list begins
with "N=", whose second byte is out of range, so it is refused before a
reader is created rather than taken for graph6 and failed on the line;
a line that is nothing but its order byte, the graph of order 0 or 1,
still passes, with or without the newline sf_fgets() leaves on it.

The predicate table gains "N=3" and "@\n", the adjacency list moves from
the read-failure cases to the r
[…]
