+++ c/planarityApp/planarityMenu.c
@@ -61,8 +61,8 @@ int menu(void)
             gp_Message("%s", GetAlgorithmSpecifiers());
 
             gp_Message(
-                "X. Transform single graph in supported file to .g6, adjacency list, or adjacency matrix\n"
-                "T. Perform an algorithm test on all graphs in .g6, .s6 or .inc.s6 input file\n"
+                "X. Transform a single graph in an input file to a supported output format\n"
+                "T. Perform an algorithm test on all graphs in an input file\n"
                 "H. Help message for command line version\n"
                 "R. Reconfigure options\n"
                 "Q. Quit\n");
