### 2026-09-23 c3327bd
feat: add GraphML reader
