+++ c/graphLib/io/graphML-reader.c
@@ -1,0 +1,1517 @@
+/*
+Copyright (c) 1997-2026, John M. Boyer
+All rights reserved.
+See the LICENSE.TXT file for licensing information.
+*/
+
+#include <ctype.h>
+#include <limits.h>
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+
+#include "graphIO.h"
+#include "strOrFile.h"
+
+#include "../graph.private.h"
+
+#define GRAPHML_MAX_ATTRIBUTES 8
+
+typedef enum
+{
+    GRAPHML_ENCODING_UTF8,
+    GRAPHML_ENCODING_ISO_8859_1
+} GraphMLEncoding;
+
+typedef struct
+{
+    char const *name;
+    int required;
+    char const *expectedValue;
+} GraphMLAttributeSpec;
+
+typedef struct
+{
+    unsigned int seen;
+    char values[GRAPHML_MAX_ATTRIBUTES][MAXLINE + 1];
+} GraphMLAttributeValues;
+
+typedef struct
+{
+    int keyPresent;
+    int keyDefault;
+} GraphMLKeyState;
+
+typedef struct
+{
+    int directedDefault;
+    int expectedNodes;
+    int expectedEdges;
+} GraphMLGraphState;
+
+int _ReadGraphMLGraph(graphP theGraph, strOrFileP inputContainer, int readGraphElemOnly);
+
+static int _ReadGraphMLStartsWith(graphP theGraph, strOrFileP inputContainer,
+                                  int *pLineNum, char const *expected, int *pMatches);
+static int _ReadGraphMLConsume(graphP theGraph, strOrFileP inputContainer,
+                               int *pLineNum, char const *expected);
+static int _ReadGraphMLSkipWhitespace(graphP theGraph, strOrFileP inputContainer,
+                                      int *pLineNum, int requireWhitespace);
+static int _ReadGraphMLSkipComment(graphP theGraph, strOrFileP inputContainer,
+                                   int *pLineNum);
+static int _ReadGraphMLSkipWhitespaceAndComments(graphP theGraph,
+                                                  strOrFileP inputContainer,
+                                                  int *pLineNum);
+static int _ReadGraphMLReadText(graphP theGraph, strOrFileP inputContainer,
+                                int *pLineNum, GraphMLEncoding encoding,
+                                int delimiter, char *buffer, size_t bufferSize);
+static int _ReadGraphMLReadName(graphP theGraph, strOrFileP inputContainer,
+                                int *pLineNum, char *buffer, size_t bufferSize);
+static int _ReadGraphMLAttribute(graphP theGraph, strOrFileP inputContainer,
+                                 int *pLineNum, GraphMLEncoding encoding,
+                                 GraphMLAttributeSpec const *specs, size_t specCount,
+                                 GraphMLAttributeValues *values);
+static int _ReadGraphMLAttributeList(graphP theGraph, strOrFileP inputContainer,
+                                     int *pLineNum, GraphMLEncoding encoding,
+                                     GraphMLAttributeSpec const *specs, size_t specCount,
+                                     GraphMLAttributeValues *values, int xmlDeclaration,
+                                     int allowEmptyTag, int *pEmptyTag);
+static int _ReadGraphMLTrim(graphP theGraph, strOrFileP inputContainer,
+                            int *pLineNum, char *value);
+static int _ReadGraphMLValidateID(graphP theGraph, strOrFileP inputContainer,
+                                  int *pLineNum, char *value);
+static int _ReadGraphMLParsePositiveInteger(graphP theGraph,
+                                             strOrFileP inputContainer,
+                                             int *pLineNum, char const *value,
+                                             int *pParsedValue);
+static int _ReadGraphMLParseCanonicalID(graphP theGraph,
[…]
