+++ c/planarityApp/planarityCommandLine.c
@@ -55,6 +55,12 @@ int runHighByteRoundTripTest(void);
 int runCapacityLimitTests(void);
 int runGraphMLWriteTest(char const *inputFileName, char const *expectedOutputFileName);
 int runBasicGraphMLWriteTest(void);
+int runGraphMLReadWriteTest(char const *inputFileName);
+int runBasicGraphMLReadTest(void);
+char *copyGraphMLTestString(char const *graphMLStr);
+int runGraphMLAcceptTest(char const *graphMLStr, int zeroBased, int directed);
+int runGraphMLRejectTest(char const *graphMLStr);
+int runGraphMLReaderStringTests(void);
 int runGraphMLTests(void);
 
 /****************************************************************************
@@ -3051,6 +3057,234 @@ int runBasicGraphMLWriteTest(void)
     return OK;
 }
 
+int runGraphMLReadWriteTest(char const *inputFileName)
+{
+    graphP G = gp_New();
+    char *actualOutput = NULL;
+    int Result = OK;
+
+    if (G == NULL)
+        return NOTOK;
+
+    if (gp_Read(G, inputFileName) != OK ||
+        gp_WriteToString(G, &actualOutput, WRITE_GRAPHML) != OK ||
+        actualOutput == NULL ||
+        TextFileMatchesString(inputFileName, actualOutput) != TRUE)
+        Result = NOTOK;
+
+    if (actualOutput != NULL)
+        free(actualOutput);
+    gp_Free(&G);
+
+    return Result;
+}
+
+int runBasicGraphMLReadTest(void)
+{
+    if (runGraphMLReadWriteTest("Digraph.transposeTest.graphml") != OK ||
+        runGraphMLReadWriteTest("Digraph.transposeTest.0-based.graphml") != OK)
+        return NOTOK;
+
+    return OK;
+}
+
+char *copyGraphMLTestString(char const *graphMLStr)
+{
+    char *copy = NULL;
+    size_t length = 0;
+
+    if (graphMLStr == NULL)
+        return NULL;
+
+    length = strlen(graphMLStr) + 1;
+    copy = (char *)malloc(length);
+    if (copy != NULL)
+        memcpy(copy, graphMLStr, length);
+
+    return copy;
+}
+
+int runGraphMLAcceptTest(char const *graphMLStr, int zeroBased, int directed)
+{
+    graphP G = gp_New();
+    char *inputCopy = NULL;
+    char *actualOutput = NULL;
+    int Result = OK;
+    int source = NIL;
+    int target = NIL;
+    int edge = NIL;
+
+    if (G == NULL || (inputCopy = copyGraphMLTestString(graphMLStr)) == NULL)
+        Result = NOTOK;
+
+    if (Result == OK && gp_ReadFromString(G, inputCopy) != OK)
+        Result = NOTOK;
+
+    if (Result == OK)
+    {
+        source = gp_LowerBoundVertexStorage(G);
+        target = source + 1;
+        edge = gp_FindEdge(G, source, target);
+
+        if (gp_GetN(G) != 2 || gp_GetM(G) != 1 || edge == NIL ||
+            (directed && gp_GetDirection(G, edge) != EDGEFLAG_DIRECTION_OUTONLY) ||
+            (!directed && gp_GetDirection(G, edge) != 0))
+            Result = NOTOK;
+    }
+
+    if (Result == OK &&
+        (gp_WriteToString(G, &actualOutput, WRITE_GRAPHML) != OK || actualOutput == NULL))
+        Result = NOTOK;
+
+    if (Result == OK &&
+        ((zeroBased && strstr(actualOutput, "<data key=\"graphflags_zerobasedio\">true</data>") == NULL) ||
+         (!zeroBased && strstr(actualOutput, "graphflags_zerobasedio") != NULL)))
[…]
