### 2026-09-23 bce3b69
Write sparse6 and incremental sparse6, byte for byte as nauty does
The sparse6 write iterator mirrors the graph6 writer's shape and adds
the change batch settled in #38: s6_StoreGraphChange() records the
deletion of an edge in use, by its endpoints, or the addition of an
edge between two real vertices that the graph lacks, and refuses loops,
parallel edges and a pair stored twice in one batch; s6_WriteGraph()
writes the whole graph as a ':' line when nothing is stored, and
otherwise the batch as a ';' line, which it then applies to the graph.
A batch is relative to the graph written last, so a
[…]
