+++ c/graphLib/graphLib.h
@@ -30,6 +30,7 @@ extern "C"
 #include "io/gp-read-iterator.h"
 #include "io/g6-write-iterator.h"
 #include "io/s6-read-iterator.h"
+#include "io/s6-write-iterator.h"
 
 // Depth-first search public API methods and definitions
 #include "graphDFSUtils.h"
