### 2026-09-25 c270c4f
Make edge storage compaction private
_CompactEdgeStorage() moves edges through gp_DeleteEdge() and
gp_DynamicAddEdge(), and an extension that overloads deletion clears the
data it keeps per edge, which insertion does not restore. As a public
method it would have to be an overloadable function in the function
table, so for now only the sparse6 reader and writer call it.
