### 2026-09-25 90de5fe
Move the sparse6 tests into planarity-test/testSparse6.c
The sparse6 read, write and edge storage compaction tests, and the
algorithm tests on all graphs from sparse6 input, move out of
planarityCommandLine.c unchanged; helpers only they use become static.
runGeneralReadIteratorTests() stays, since it covers both formats.
