+++ c/planarityApp/planarity-test/testSparse6.c
@@ -1,0 +1,1353 @@
+/*
+Copyright (c) 1997-2026, John M. Boyer
+All rights reserved.
+See the LICENSE.TXT file for licensing information.
+*/
+
+#include <stdlib.h>
+#include <string.h>
+
+#include "../planarity.h"
+#include "../../graphLib/io/strOrFile.h"
+
+/* Library-private function the edge storage tests call directly */
+extern int _CompactEdgeStorage(graphP theGraph);
+
+/* Called from planarityCommandLine.c */
+int runSparse6ReadTests(void);
+int runSparse6WriteTests(void);
+int runSparse6TestAllGraphsTests(void);
+char *copySparse6TestString(char const *s6Str);
+
+/* Defined in planarityCommandLine.c */
+int runTestAllGraphsTest(char const *commandString, char const *infileName, char const *expectedValidationStr);
+int runGraphTransformationTest(char const *command, char const *infileName, int inputInMemFlag);
+
+static int compareSparse6Output(char const *ours, char const *expected, char const *what);
+static int runSparse6WriteLineTest(char const *g6Str, char const *expectedLine);
+static int runSparse6WriteRoundTripTest(char const *g6FileName, char const *s6FileName, int incremental);
+static int runSparse6WriteContractTests(void);
+static int runCompactEdgeStorageTests(void);
+static int runSparse6LockstepTest(char const *g6FileName, char const *s6FileName, int inputInMemFlag, int expectedNumGraphs);
+static int runSparse6AcceptTest(char const *s6Str, char const *expectedG6Line);
+static int runSparse6AcceptEdgesTest(char const *s6Str, int order, int const edges[][2], int numEdges);
+static int runSparse6RejectTest(char const *s6Str);
+
+/****************************************************************************
+ compareSparse6Output()
+
+ Compares the output of the sparse6 writer, which begins with the header,
+ with the expected lines, which may or may not, and reports the first line
+ that differs. Returns OK when every line matches, NOTOK otherwise.
+ ****************************************************************************/
+
+static int compareSparse6Output(char const *ours, char const *expected, char const *what)
+{
+    char const *s6Header = ">>sparse6<<";
+    char const *p = ours;
+    char const *q = expected;
+    int lineNum = 1;
+
+    if (ours == NULL || expected == NULL)
+    {
+        gp_ErrorMessage("No sparse6 output to compare for %s.", what);
+        return NOTOK;
+    }
+
+    if (strncmp(p, s6Header, strlen(s6Header)) != 0)
+    {
+        gp_ErrorMessage("The sparse6 output for %s does not begin with the header.", what);
+        return NOTOK;
+    }
+    p += strlen(s6Header);
+
+    if (strncmp(q, s6Header, strlen(s6Header)) == 0)
+        q += strlen(s6Header);
+
+    while (*p != '\0' || *q != '\0')
+    {
+        size_t ourLen = strcspn(p, "\r\n");
+        size_t expLen = strcspn(q, "\r\n");
+
+        if (ourLen != expLen || strncmp(p, q, ourLen) != 0)
+        {
+            gp_ErrorMessage("Line %d of the sparse6 output for %s is \"%.*s\" "
+                            "rather than \"%.*s\".",
+                            lineNum, what, (int)ourLen, p, (int)expLen, q);
+            return NOTOK;
+        }
+
+        p += ourLen;
+        q += expLen;
+        while (*p == '\r' || *p == '\n')
+            p++;
+        while (*q == '\r' || *q == '\n')
+            q++;
+        lineNum++;
+    }
+
+    return OK;
+}
+
[…]
