+++ c/planarityApp/planarityCommandLine.c
@@ -39,18 +39,9 @@ int runHideRestoreTest(graphP theGraph);
 int runIdentifyContractTest(graphP theGraph);
 int runSparse6ReadTests(void);
 int runGeneralReadIteratorTests(void);
-int runSparse6LockstepTest(char const *g6FileName, char const *s6FileName, int inputInMemFlag, int expectedNumGraphs);
-int runSparse6AcceptTest(char const *s6Str, char const *expectedG6Line);
-int runSparse6AcceptEdgesTest(char const *s6Str, int order, int const edges[][2], int numEdges);
-int runSparse6RejectTest(char const *s6Str);
 char *copySparse6TestString(char const *s6Str);
+int runSparse6TestAllGraphsTests(void);
 int runSparse6WriteTests(void);
-int runSparse6WriteLineTest(char const *g6Str, char const *expectedLine);
-int runSparse6WriteRoundTripTest(char const *g6FileName, char const *s6FileName, int incremental);
-int runSparse6WriteContractTests(void);
-int runCompactEdgeStorageTests(void);
-extern int _CompactEdgeStorage(graphP theGraph);
-int compareSparse6Output(char const *ours, char const *expected, char const *what);
 int testDirectedDFS(void);
 int testPetersenDigraph(void);
 int testDigraphTranspose(void);
@@ -1116,1269 +1107,6 @@ int runGeneralReadIteratorTests(void)
     return Result;
 }
 
-/****************************************************************************
- compareSparse6Output()
-
- Compares the output of the sparse6 writer, which begins with the header,
- with the expected lines, which may or may not, and reports the first line
- that differs. Returns OK when every line matches, NOTOK otherwise.
- ****************************************************************************/
-
-int compareSparse6Output(char const *ours, char const *expected, char const *what)
-{
-    char const *s6Header = ">>sparse6<<";
-    char const *p = ours;
-    char const *q = expected;
-    int lineNum = 1;
-
-    if (ours == NULL || expected == NULL)
-    {
-        gp_ErrorMessage("No sparse6 output to compare for %s.", what);
-        return NOTOK;
-    }
-
-    if (strncmp(p, s6Header, strlen(s6Header)) != 0)
-    {
-        gp_ErrorMessage("The sparse6 output for %s does not begin with the header.", what);
-        return NOTOK;
-    }
-    p += strlen(s6Header);
-
-    if (strncmp(q, s6Header, strlen(s6Header)) == 0)
-        q += strlen(s6Header);
-
-    while (*p != '\0' || *q != '\0')
-    {
-        size_t ourLen = strcspn(p, "\r\n");
-        size_t expLen = strcspn(q, "\r\n");
-
-        if (ourLen != expLen || strncmp(p, q, ourLen) != 0)
-        {
-            gp_ErrorMessage("Line %d of the sparse6 output for %s is \"%.*s\" "
-                            "rather than \"%.*s\".",
-                            lineNum, what, (int)ourLen, p, (int)expLen, q);
-            return NOTOK;
-        }
-
-        p += ourLen;
-        q += expLen;
-        while (*p == '\r' || *p == '\n')
-            p++;
-        while (*q == '\r' || *q == '\n')
-            q++;
-        lineNum++;
-    }
-
-    return OK;
-}
-
-/****************************************************************************
- runSparse6WriteLineTest()
-
- Reads one graph from the graph6 string and checks that the sparse6 writer
- produces exactly the expected line for it, which is the line nauty writes.
- ****************************************************************************/
-
[…]
