### 2026-09-25 2624cc5
Read and write sparse6 of any order a graph can have
The reader parses the eight-byte order encoding, for orders from 258048
on, refusing an order that does not fit an int before converting it, and
leaves the vertex capacity check to the graph, which refuses more than
178,956,970 vertices before allocating anything. The writer writes the
eight-byte encoding and no longer refuses orders over 100,000. It sends
a line out through a fixed 64 KiB buffer instead of building it whole,
since its old line buffer assumed at most 18 bits per vertex and a line
near the largest order can exceed what one sf
[…]
