+++ c/graphLib/io/s6-write-iterator.c
@@ -345,12 +345,6 @@ int _s6_InitWriter(S6WriteIteratorP theS6WriteIterator)
     char const *s6Header = ">>sparse6<<";
     int order = gp_GetN(theS6WriteIterator->currGraph);
 
-    if (order > 100000)
-    {
-        gp_ErrorMessage("Graphs of order n > 100000 are not supported at this time.");
-        return NOTOK;
-    }
-
     if (sf_fputs(s6Header, theS6WriteIterator->outputContainer) < 0)
     {
         gp_ErrorMessage("Unable to initialize writer due to failure to fputs "
@@ -761,16 +755,45 @@ int _s6_ComparePairs(void const *a, void const *b)
     return 0;
 }
 
-// Appends the low width bits of value to the bit stream, writing each
-// completed group of six bits as a byte in the range 63 to 126
+// A line is written out through a buffer of this many bytes, so that a line
+// of any length, which for a graph near the largest order can exceed what
+// one sf_fputs() accepts, needs no more memory than that
+#define S6_LINE_CHUNK_SIZE 65536
+
 typedef struct
 {
     char *out;
     size_t pos;
     unsigned long long acc;
     int numBits;
+    strOrFileP outputContainer;
+    int failed;
 } s6BitWriter;
 
+// Writes the bytes gathered so far to the output container, and remembers
+// a failure so that the caller can report it once the line is done
+static void _s6_FlushBytes(s6BitWriter *writer)
+{
+    if (writer->pos == 0)
+        return;
+
+    writer->out[writer->pos] = '\0';
+    if (!writer->failed && sf_fputs(writer->out, writer->outputContainer) < 0)
+        writer->failed = TRUE;
+    writer->pos = 0;
+}
+
+// Appends one byte, keeping room in the buffer for the null terminator
+static void _s6_PutByte(s6BitWriter *writer, char byte)
+{
+    if (writer->pos == S6_LINE_CHUNK_SIZE - 1)
+        _s6_FlushBytes(writer);
+
+    writer->out[writer->pos++] = byte;
+}
+
+// Appends the low width bits of value to the bit stream, writing each
+// completed group of six bits as a byte in the range 63 to 126
 static void _s6_PutBits(s6BitWriter *writer, unsigned int value, int width)
 {
     writer->acc = (writer->acc << width) | (value & ((1ULL << width) - 1));
@@ -779,7 +802,7 @@ static void _s6_PutBits(s6BitWriter *writer, unsigned int value, int width)
     while (writer->numBits >= 6)
     {
         writer->numBits -= 6;
-        writer->out[writer->pos++] = (char)(((writer->acc >> writer->numBits) & 63) + 63);
+        _s6_PutByte(writer, (char)(((writer->acc >> writer->numBits) & 63) + 63));
     }
 
     writer->acc &= (1ULL << writer->numBits) - 1;
@@ -808,65 +831,64 @@ static void _s6_PutBits(s6BitWriter *writer, unsigned int value, int width)
  which keeps a third int per pair, be encoded in place.
 
  The line, including its terminator, is written to the output
- container. Returns OK on success, NOTOK otherwise.
+ container in pieces of at most S6_LINE_CHUNK_SIZE bytes, so its length
+ is limited only by the order and the number of edges. Returns OK on
+ success, NOTOK otherwise.
  ********************************************************************/
 
 int _s6_EncodeLine(S6WriteIteratorP theS6WriteIterator, char lineChar, int const *pairs, int numPairs, int stride)
 {
     const int order = theS6WriteIterator->order;
[…]
