+++ c/graphLib/graphDFSUtils.c
@@ -102,6 +102,11 @@ int gp_DepthFirstSearch(graphP theGraph)
 
     if (theGraph == NULL)
         return NOTOK;
+    if (theGraph->graphFlags & GRAPHFLAGS_PARALLELEDGEDETECTED)
+    {
+        gp_ErrorMessage("Loop edges were previously added to the graph. See gp_DeleteParallelEdges().");
+        return NOTOK;
+    }
 
     if (gp_GetGraphFlags(theGraph) & GRAPHFLAGS_DFSNUMBERED)
         return OK;
@@ -550,6 +555,12 @@ int gp_ComputeLowpoints(graphP theGraph)
     if (theGraph == NULL)
         return NOTOK;
 
+    if (theGraph->graphFlags & GRAPHFLAGS_PARALLELEDGEDETECTED)
+    {
+        gp_ErrorMessage("Loop edges were previously added to the graph. See gp_DeleteParallelEdges().");
+        return NOTOK;
+    }
+
     if (gp_GetGraphFlags(theGraph) & GRAPHFLAGS_DIRECTEDEDGEDETECTED)
     {
         gp_ErrorMessage("gp_ComputeLowpoints() does not support directed graphs.");
@@ -676,6 +687,12 @@ int gp_ComputeLeastAncestors(graphP theGraph)
 
     if (theGraph == NULL)
         return NOTOK;
+
+    if (theGraph->graphFlags & GRAPHFLAGS_PARALLELEDGEDETECTED)
+    {
+        gp_ErrorMessage("Loop edges were previously added to the graph. See gp_DeleteParallelEdges().");
+        return NOTOK;
+    }
 
     if (gp_GetGraphFlags(theGraph) & GRAPHFLAGS_DIRECTEDEDGEDETECTED)
     {
