+++ c/planarityApp/planarityCommandLine.c
@@ -16,6 +16,7 @@ See the LICENSE.TXT file for licensing information.
 #include <unistd.h>
 #endif
 
+int runParallelEdgeTests();
 int runQuickRegressionTests(int argc, char *argv[]);
 int callRandomGraphs(int argc, char *argv[]);
 int callSpecificGraph(int argc, char *argv[]);
@@ -208,7 +209,7 @@ int runQuickRegressionTests(int argc, char *argv[])
 
     if (!getcwd(origDir, 2 * MAXLINE))
         return NOTOK;
-
+   
     // Preserve original behavior before the samplesDir command-line parameter was available
     if (strcmp(samplesDir, "samples") == 0)
     {
@@ -238,8 +239,9 @@ int runQuickRegressionTests(int argc, char *argv[])
             return OK;
         }
     }
-
-    if (runSpecificGraphTests() != OK)
+     if (runParallelEdgeTests() != OK) 
+        retVal =  NOTOK;
+    else if (runSpecificGraphTests() != OK)
         retVal = NOTOK;
     else if (runRandomGraphsTests() != OK)
         retVal = NOTOK;
@@ -2193,4 +2195,46 @@ int runGraphMLTests(void)
         gp_Message("Finished GraphML Tests.\n");
 
     return Result;
+}
+int runParallelEdgeTests(void)
+{
+    graphP G = gp_New();
+    if (G == NULL) return NOTOK;
+    
+    graphP G1 = NULL;
+    char filename[] = "Petersen-with-parallel-edges.txt";
+    unsigned quietModeCache;
+
+    if (gp_Read(G, filename) != OK) return NOTOK;
+
+    if (!(G->graphFlags & GRAPHFLAGS_PARALLELEDGEDETECTED)) return NOTOK;
+    if (gp_GetM(G) != 60) return NOTOK;
+
+    G1 = gp_New();
+    if (G1 == NULL) return NOTOK;
+
+    if (gp_CopyGraph(G1, G) != OK) return NOTOK;
+
+    if (!(G1->graphFlags & GRAPHFLAGS_PARALLELEDGEDETECTED)) return NOTOK;
+    if (gp_GetM(G1) != 60) return NOTOK;
+    quietModeCache = gp_GetQuietMode();
+    gp_SetQuietMode(QUIETMODE_ALL);
+
+    if (gp_DepthFirstSearch(G1) == OK) return NOTOK;
+    if (gp_ComputeLowpoints(G1) == OK) return NOTOK;
+    if (gp_ComputeLeastAncestors(G1) == OK) return NOTOK;
+    if (gp_Embed(G1, EMBEDFLAGS_PLANAR) == OK) return NOTOK;
+    gp_SetQuietMode(quietModeCache);
+
+    if (gp_DeleteParallelEdges(G1) != OK) return NOTOK;   
+    if (G1->graphFlags & GRAPHFLAGS_PARALLELEDGEDETECTED) return NOTOK;
+    if (gp_GetM(G1) != 15) return NOTOK;
+    if (gp_Embed(G1, EMBEDFLAGS_PLANAR) != NONEMBEDDABLE) return NOTOK;
+    
+    if (gp_TestEmbedResultIntegrity(G1, G, EMBEDFLAGS_PLANAR) != OK) return NOTOK;
+
+    gp_Free(&G);
+    gp_Free(&G1);
+
+    return OK;
 }
