### 2026-09-26 4143fcf
Implement parallel edge deletion, update edge capacity, and add regression tests
