+++ c/planarityApp/planarityCommandLine.c
@@ -3073,23 +3073,20 @@ int runParallelEdgeTests(void)
 {
     graphP G = gp_New();
     if (G == NULL) return NOTOK;
-    
     graphP G1 = NULL;
-    char filename[] = "Petersen-with-parallel-edges.txt";
     unsigned quietModeCache;
 
-    if (gp_Read(G, filename) != OK) return NOTOK;
+    if (gp_Read(G, "Petersen-with-parallel-edges.txt") != OK && 
+        gp_Read(G, "c/samples/Petersen-with-parallel-edges.txt") != OK) {
+        return NOTOK;
+    }
 
     if (!(G->graphFlags & GRAPHFLAGS_PARALLELEDGEDETECTED)) return NOTOK;
     if (gp_GetM(G) != 60) return NOTOK;
 
-    G1 = gp_New();
+    G1 = gp_DupGraph(G);
     if (G1 == NULL) return NOTOK;
 
-    if (gp_CopyGraph(G1, G) != OK) return NOTOK;
-
-    if (!(G1->graphFlags & GRAPHFLAGS_PARALLELEDGEDETECTED)) return NOTOK;
-    if (gp_GetM(G1) != 60) return NOTOK;
     quietModeCache = gp_GetQuietMode();
     gp_SetQuietMode(QUIETMODE_ALL);
 
@@ -3097,15 +3094,21 @@ int runParallelEdgeTests(void)
     if (gp_ComputeLowpoints(G1) == OK) return NOTOK;
     if (gp_ComputeLeastAncestors(G1) == OK) return NOTOK;
     if (gp_Embed(G1, EMBEDFLAGS_PLANAR) == OK) return NOTOK;
+
     gp_SetQuietMode(quietModeCache);
 
-    if (gp_DeleteParallelEdges(G1) != OK) return NOTOK;   
+    // Delete parallel edges and verify exact 15-edge state 
+    if (gp_DeleteParallelEdges(G1) != OK) return NOTOK;
     if (G1->graphFlags & GRAPHFLAGS_PARALLELEDGEDETECTED) return NOTOK;
     if (gp_GetM(G1) != 15) return NOTOK;
-    if (gp_Embed(G1, EMBEDFLAGS_PLANAR) != NONEMBEDDABLE) return NOTOK;
     
-    if (gp_TestEmbedResultIntegrity(G1, G, EMBEDFLAGS_PLANAR) != OK) return NOTOK;
+    //  Verify embedding and structural integrity of the cleaned graph 
+    if (gp_Embed(G1, EMBEDFLAGS_PLANAR) != NONEMBEDDABLE) return NOTOK;
 
+    if (gp_TestEmbedResultIntegrity(G1, G, NONEMBEDDABLE) != NONEMBEDDABLE) {
+        return NOTOK;
+    }
+    
     gp_Free(&G);
     gp_Free(&G1);
 
