+++ c/samples/Petersen-with-parallel-edges.txt
@@ -1,11 +1,11 @@
 N=10
-1: 2 2 2 2 6 6 6 6 5 5 5 5 0
-2: 3 3 3 3 7 7 7 7 1 1 1 1 0
-3: 4 4 4 4 8 8 8 8 2 2 2 2 0
-4: 5 5 5 5 9 9 9 9 3 3 3 3 0
-5: 1 1 1 1 10 10 10 10 4 4 4 4 0
-6: 1 1 1 1 8 8 8 8 9 9 9 9 0
-7: 2 2 2 2 9 9 9 9 10 10 10 10 0
-8: 3 3 3 3 10 10 10 10 6 6 6 6 0
-9: 4 4 4 4 6 6 6 6 7 7 7 7 0
-10: 5 5 5 5 7 7 7 7 8 8 8 8 0
+1: 2 2 2 2 5 5 5 5 6 6 6 6 0
+2: 3 3 3 3 7 7 7 7 0
+3: 4 4 4 4 8 8 8 8 0
+4: 5 5 5 5 9 9 9 9 0
+5: 10 10 10 10 0
+6: 8 8 8 8 9 9 9 9 0
+7: 9 9 9 9 10 10 10 10 0
+8: 10 10 10 10 0
+9: 0
+10: 0
