+++ c/graphLib/745A.cpp
@@ -1,25 +1,0 @@
-#include <bits/stdc++.h>
-using namespace std;
-
-
-
-void solve(){
-    
-  
-}
-
-
-
-
-
-
-int main() {
-    ios::sync_with_stdio(false);
-    cin.tie(nullptr);
-    int t; cin >> t;
-    while(t--){
-        solve();
-    }
-  
-    return 0;
-}
