### 2026-09-26 cbff65e
Give quiet treatment to an intentional NOTOK test case
