### 2026-09-26 f55501a
Minor text layout changes in planarity -test
