### 2026-09-26 d1101da
Ask codespell to ignore a spelling error that isn't one
