### 2026-09-26 eb1722d
Match exact type of source data to clear MSVC two warnings
