### 2026-09-26 2766673
Codespell seems to want lowercase to match dictionary entry, even though error is shown in uppercase.
