+++ c/planarityApp/planarity-test/testSparse6.c
@@ -1352,7 +1352,7 @@ static int runSparse6LargeOrderTests(void)
         {":~}~~_??^n~fv~n\n", 258047, 2, {{0, 1}, {258045, 258046}}},
         {":~~???~??~^~_??N\n", 258048, 1, {{0, 258047}, {0, 0}}},
         {":~~??@???_??^~~_??^\n", 262144, 2, {{0, 1}, {3, 262142}}},
-        {":~~??@HN_qRvo??THN]\n", 300000, 2, {{5, 299999}, {299998, 299999}}}, // codespell:ignore THN
+        {":~~??@HN_qRvo??THN]\n", 300000, 2, {{5, 299999}, {299998, 299999}}}, // codespell:ignore thn
     };
     int Result = OK;
 
